clean & split fem_ele_std, fem_ele_vec
