+++ FEM/FileIO/vtk.cpp
@@ -18,10 +18,14 @@
 #include <sys/types.h>
 #endif
 
-#include "StringTools.h"
-#include "Output.h"
-#include "fem_ele_std.h"  // for element velocity
 #include "makros.h"
+#include "StringTools.h"
+
+#include "mathlib.h"
+
+#include "ElementValue.h"
+#include "Output.h"
+#include "fem_ele_std.h"
 #include "rf_mmp_new.h"
 
 using namespace std;
