+++ FEM/fem_ele_std.cpp
@@ -17,16 +17,18 @@
 #ifdef NEW_EQS
 #include "equation_class.h"
 #endif
-#include "eos.h"
 #include "mathlib.h"
 #if defined(USE_PETSC)
 #include "PETSC/PETScLinearSolver.h"
 #endif
+#include "SparseMatrixDOK.h"
+
+#include "ElementValue.h"
+#include "eos.h"
 #include "rfmat_cp.h"
 #include "rf_mmp_new.h"
 #include "rf_msp_new.h"
 #include "rf_pcs_dm.h"
-#include "SparseMatrixDOK.h"
 
 #ifdef NEW_EQS
 using Math_Group::CSparseMatrix;
@@ -1148,205 +1150,6 @@ double CFiniteElementStd::CalCoefMass()
    FEMLib-Method:
    Task: Calculate material coefficient for mass matrix
    Programing:
-   02/2007 WW Multi-phase flow
-   05/2008 WW Generalization
-**************************************************************************/
-double CFiniteElementStd::CalCoefMass2(int dof_index)
-{
-	int Index = MeshElement->GetIndex();
-	double val = 0.0;
-	double expfactor = 0.0;
-	double dens_arg[3];               // 08.05.2008 WW
-	double pert = sqrt(DBL_EPSILON);  // 15.08.2011. WW
-
-	bool diffusion = false;  // 08.05.2008 WW
-
-	if (MediaProp->heat_diffusion_model == 1 && cpl_pcs) diffusion = true;
-	dens_arg[1] = 293.15;
-	//
-	if (pcs->m_num->ele_mass_lumping) ComputeShapefct(1);
-	// CB_merge_0513 in case of het K, store local K in permeability_tensor
-	double* tensor = NULL;
-	tensor = MediaProp->PermeabilityTensor(Index);
-	MediaProp->local_permeability = tensor[0];
-
-	switch (dof_index)
-	{
-		case 0:
-			PG = interpolate(NodalVal1);  // Capillary pressure
-			dens_arg[0] = PG;             // Should be P_w in some cases
-			if (diffusion)
-			{
-				TG = interpolate(NodalValC1) + T_KILVIN_ZERO;
-				dens_arg[1] = TG;
-			}
-			Sw = MediaProp->SaturationCapillaryPressureFunction(PG);
-			rhow = FluidProp->Density(dens_arg);
-			dSdp = MediaProp->PressureSaturationDependency(
-			    Sw, true);  // JT: dSdp now returns actual sign (i.e. <0)
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			// Storativity   28.05.2008
-			// val = MediaProp->StorageFunction(Index,unit,pcs->m_num->time_theta)
-			// *Sw;
-			// Fluid compressibility
-			// val += poro  *Sw* FluidProp->drho_dp / rhow;
-			if (SolidProp)
-			{
-				if (SolidProp->Ks > MKleinsteZahl)  // Storativity WX:28.05.2008
-					val -= Sw * (SolidProp->biot_const - poro) / SolidProp->Ks *
-					       Sw;
-			}
-			// Fluid compressibility
-			if (fabs(FluidProp->drho_dp) > MKleinsteZahl)
-				val -= poro * Sw * FluidProp->drho_dp;
-			val += poro * dSdp;  // WX:04.2013 val = poro * dSdp;
-
-			// Coupled (T)
-			if (diffusion)
-			{
-				// Water vapour pressure
-				expfactor = COMP_MOL_MASS_WATER / (rhow * GAS_CONSTANT * TG);
-				rho_gw = FluidProp->vaporDensity(TG) * exp(-PG * expfactor);
-				//
-				val -= poro * dSdp * rho_gw / rhow;
-				//
-				val -= (1.0 - Sw) * poro * COMP_MOL_MASS_WATER * rho_gw /
-				       (rhow * GAS_CONSTANT * TG * rhow);
-			}
-			break;
-		case 1:  // 01
-			val = 0.0;
-			// WX:05.2012 Storgae
-			PG = interpolate(NodalVal1);
-			Sw = MediaProp->SaturationCapillaryPressureFunction(PG);
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			if (SolidProp)
-			{
-				if (SolidProp->Ks > MKleinsteZahl)
-					// WX:11.2012
-					val += Sw * (SolidProp->biot_const - poro) / SolidProp->Ks;
-			}
-			// WX:05.2012 Compressibility
-			if (fabs(FluidProp->drho_dp) > MKleinsteZahl)
-				val += poro * Sw * FluidProp->drho_dp;
-
-			break;
-		case 2:  //
-			// (1-S)n(d rhop_c/d p_c)
-			PG2 = interpolate(NodalVal_p2);
-			dens_arg[0] = PG2;  // 28.05.2008. WW
-			val = 0.;           // 28.05.2008. WW
-			if (diffusion)      // 28.05.2008. WW
-			{
-				val = (1.0 - Sw) * COMP_MOL_MASS_WATER * rho_gw /
-				      (rhow * GAS_CONSTANT * TG * rhow);
-				p_gw = rho_gw * GAS_CONSTANT * TG / COMP_MOL_MASS_WATER;
-				dens_arg[0] -= p_gw;
-				dens_arg[1] = TG;
-			}
-			rho_ga = GasProp->Density(dens_arg);  // 28.05.2008. WW
-			val -= rho_ga * dSdp / rhow;
-			val *= poro;
-			// WX:11.2012.storage
-			if (SolidProp)
-			{
-				if (SolidProp->Ks > MKleinsteZahl)
-					val -= rho_ga / rhow *
-					       ((1 - Sw) * (SolidProp->biot_const - poro) /
-					        SolidProp->Ks * Sw);
-			}
-
-			break;
-		case 3:  //
-			// Approximation of d dens_g/dp_g 16.08.2011. WW
-			dens_arg[0] = PG2 + pert;
-			if (diffusion) dens_arg[1] = TG;
-			/// d dens_g/dp_g:
-			if (GasProp->density_model ==
-			    2)  // dens_g/dp_g = drho_dp 02.2012. WX
-				val = (1.0 - Sw) * poro * GasProp->rho_0 * GasProp->drho_dp /
-				      rhow;
-			else
-				val = (1.0 - Sw) * poro *
-				      (GasProp->Density(dens_arg) - rho_ga) / (pert * rhow);
-			// Storage WX:11.2012
-			if (SolidProp)
-			{
-				if (SolidProp->Ks > MKleinsteZahl)
-					val += (SolidProp->biot_const - poro) / SolidProp->Ks *
-					       (1 - Sw) * GasProp->Density(dens_arg) / rhow;
-			}
-
-			break;
-	}
-	return val;
-}
-
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate material coefficient for mass matrix
-   Programing:
-   03/2009 PCH Multi-phase flow
-**************************************************************************/
-double CFiniteElementStd::CalCoefMassPSGLOBAL(int dof_index)
-{
-	int Index = MeshElement->GetIndex();
-	double val = 0.0, variables[3];
-	double P, T;
-	// OK411 double expfactor = 0.0;
-	// WW bool diffusion = false;                     //08.05.2008 WW
-	// WWif(MediaProp->heat_diffusion_model==273&&cpl_pcs)
-	// WW  diffusion = true;
-	//
-	if (pcs->m_num->ele_mass_lumping) ComputeShapefct(1);
-	switch (dof_index)
-	{
-		case 0:
-
-			// compressibility also for the wetting phase NB
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			Sw = 1.0 - interpolate(NodalVal_SatNW);  // Sw = 1-Snw
-			P = interpolate(NodalVal1);              // Pw
-			T = interpolate(NodalValC1);
-			variables[0] = P;
-			variables[1] = T;
-			val =
-			    poro * (Sw)*FluidProp->drhodP(variables) / FluidProp->Density();
-			//		cout << FluidProp->fluid_name << " Pressure: " << P << "
-			//Temp:
-			//" << ": drhodP: " << FluidProp->drhodP(P,T) << " density: " <<
-			// FluidProp->Density() << endl;
-			break;
-		case 1:  // Snw in the wetting equation
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			val = -poro;
-			break;
-		case 2:  // Pw in the non-wetting equation
-			Sw = 1.0 - interpolate(NodalVal_SatNW);  // Sw = 1 - Snw
-			// Pnw = Pw + Pc(Sw)
-			P = interpolate(NodalVal1) +
-			    MediaProp->CapillaryPressureFunction(Sw);
-			//      P = interpolate(NodalVal1);  // Pw
-			T = interpolate(NodalValC1);
-			variables[0] = P;
-			variables[1] = T;
-			val = poro * (1. - Sw) * GasProp->drhodP(variables) /
-			      GasProp->Density();
-
-			break;
-		case 3:  // Snw in the non-wetting equation
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			val = poro;
-			break;
-	}
-
-	return val;
-}
-
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate material coefficient for mass matrix
-   Programing:
    01/2005 WW/OK Implementation
    03/2005 WW Heat transport
    07/2005 WW Change for geometry element object
@@ -1911,848 +1714,7 @@ void CFiniteElementStd::CalCoefLaplace(bool Gravity, int ip)
 	}
 }
 
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate material coefficient for Laplacian matrix
-   Programing:
-   10/2008 PCH Implementation
-**************************************************************************/
-void CFiniteElementStd::CalCoefLaplaceMultiphase(int phase, int /*ip*/)
-{
-	int i = 0;
-	double mat_fac = 1.0;
-	double* tensor = NULL;
-	// static double Hn[9],z[9];
-	int Index = MeshElement->GetIndex();
-	double k_rel;
-	ComputeShapefct(1);  //  12.3.2007 WW
 
-	// For nodal value interpolation
-	//======================================================================
-	switch (PcsType)
-	{
-		default:
-			break;
-
-		case T:  // Two-phase flow
-			// PCH Rewriting...
-			// PCH Laplace mat_fac is accounted for two phases here.
-			// thought to be related to the reference pressure.
-			tensor = MediaProp->PermeabilityTensor(Index);
-			if (pcs->pcs_type_number == 0)
-			{
-				// PCH Laplace mat_fac is accounted for two phases here.
-				// thought to be related to the reference pressure.
-				double mat_fac = 0.0;
-
-				idxS = cpl_pcs->GetNodeValueIndex("SATURATION2");
-
-				for (i = 0; i < nnodes; i++)
-					NodalVal_Sat[i] = cpl_pcs->GetNodeValue(nodes[i], idxS + 1);
-				Sw = 1.0 - interpolate(NodalVal_Sat);
-
-				k_rel = MediaProp->PermeabilitySaturationFunction(Sw, phase);
-
-				// Note here mat_fac is += meaning adding two phases
-				mat_fac =
-				    time_unit_factor * k_rel / mfp_vector[phase]->Viscosity();
-
-				for (size_t i = 0; i < dim * dim; i++)
-					mat[i] = tensor[i] * mat_fac;
-			}
-			else if (pcs->pcs_type_number == 1)
-			{
-				int phase = pcs->pcs_type_number;
-
-				idxS = pcs->GetNodeValueIndex("SATURATION2");
-				for (i = 0; i < nnodes; i++)
-					NodalVal_Sat[i] = pcs->GetNodeValue(nodes[i], idxS + 1);
-				Sw = 1.0 - interpolate(NodalVal_Sat);
-				k_rel = MediaProp->PermeabilitySaturationFunction(Sw, phase);
-
-				// Here only the second phase accounted.
-				mat_fac =
-				    time_unit_factor * k_rel / mfp_vector[phase]->Viscosity();
-				for (size_t i = 0; i < dim * dim; i++)
-					mat[i] = tensor[i] * mat_fac;
-			}
-			break;
-	}
-}
-
-///////
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate material coefficient for Laplacian matrix of multi-phase
-      flow
-   Programing:
-   02/2007 WW Implementation
-   last modification:
-**************************************************************************/
-void CFiniteElementStd::CalCoefLaplace2(bool Gravity, int dof_index)
-{
-	double* tensor = NULL;
-	double mat_fac = 1.0, m_fac = 0.;
-	double fac_perm = 1.;  // WX: factor for Permeability as funktion of
-	                       // pressure, strain, etc ... 05.2010
-	double expfactor, D_gw, D_ga;
-	expfactor = D_gw = D_ga = 0.0;
-	double dens_arg[3];      // 08.05.2008 WW
-	bool diffusion = false;  // 08.05.2008 WW
-	if (MediaProp->heat_diffusion_model == 1 && cpl_pcs) diffusion = true;
-	//
-	dens_arg[1] = 293.15;
-	//
-	int Index = MeshElement->GetIndex();
-	// CB_merge_0513 in case of het K, store local K in permeability_tensor
-	tensor = MediaProp->PermeabilityTensor(Index);
-	MediaProp->local_permeability = tensor[0];
-	//
-	ComputeShapefct(1);  //  12.3.2007 WW
-
-	// WX: 11.05.2010
-	PG = interpolate(NodalVal1);
-	PG2 = interpolate(NodalVal_p2);
-	// WX: cal factor for permeability 11.05.2010
-	CFiniteElementStd* h_fem;
-	h_fem = this;
-
-	if (MediaProp->permeability_pressure_model > 0)  // 01.09.2011. WW
-		fac_perm = MediaProp->PermeabilityFunctionPressure(Index, PG2);
-	if (MediaProp->permeability_strain_model > 0)  // 01.09.2011. WW
-		fac_perm *= MediaProp->PermeabilityFunctionStrain(Index, nnodes, h_fem);
-	//======================================================================
-	for (size_t i = 0; i < dim * dim; i++)
-		mat[i] = 0.0;
-	switch (dof_index)
-	{
-		case 0:
-		{
-			PG = interpolate(NodalVal1);
-			Sw = MediaProp->SaturationCapillaryPressureFunction(PG);
-			//
-			tensor = MediaProp->PermeabilityTensor(Index);
-			mat_fac = MediaProp->PermeabilitySaturationFunction(Sw, 0) /
-			          FluidProp->Viscosity();
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = -tensor[i] * mat_fac * time_unit_factor *
-				         fac_perm;  // WX:05.2010
-			// For velocity caculation
-			if (!Gravity)
-			{
-				dens_arg[0] = PG;  // Shdould be Pw in some cases
-				if (diffusion)
-				{
-					TG = interpolate(NodalValC1) + T_KILVIN_ZERO;
-					dens_arg[1] = TG;
-				}
-				//
-				rhow = FluidProp->Density(dens_arg);
-				poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-				PG2 = interpolate(NodalVal_p2);
-				dens_arg[0] = PG2;
-				//
-				if (diffusion)
-				{
-					tort = MediaProp->TortuosityFunction(Index, unit,
-					                                     pcs->m_num->time_theta);
-					tort *= MediaProp->base_heat_diffusion_coefficient *
-					        (1 - Sw) * poro * pow(TG / T_KILVIN_ZERO, 1.8);
-					expfactor =
-					    COMP_MOL_MASS_WATER / (rhow * GAS_CONSTANT * TG);
-					rho_gw = FluidProp->vaporDensity(TG) * exp(-PG * expfactor);
-					p_gw = rho_gw * GAS_CONSTANT * TG / COMP_MOL_MASS_WATER;
-					dens_arg[0] -= p_gw;
-				}
-				//
-				rho_ga = GasProp->Density(dens_arg);
-				//
-				if (diffusion)
-				{
-					rho_g = rho_ga + rho_gw;
-					// 1/Mg
-					M_g = (rho_gw / COMP_MOL_MASS_WATER +
-					       rho_ga / GasProp->molar_mass) /
-					      rho_g;
-					D_gw = tort * rho_g * COMP_MOL_MASS_WATER *
-					       GasProp->molar_mass * M_g * M_g / rhow;
-					D_gw *= rho_gw / (rhow * PG2);
-					for (size_t i = 0; i < dim; i++)
-						mat[i * dim + i] -= D_gw * time_unit_factor;
-				}
-			}
-			break;
-		}
-		case 1:
-			if (Gravity)
-			{
-				PG = interpolate(NodalVal1);
-				PG2 = interpolate(NodalVal_p2);
-				Sw = MediaProp->SaturationCapillaryPressureFunction(PG);
-				dens_arg[0] = PG;  // Shdould be Pw in some cases
-				if (diffusion)
-				{
-					TG = interpolate(NodalValC1) + T_KILVIN_ZERO;
-					dens_arg[1] = TG;
-				}
-				// Liquid density
-				rhow = FluidProp->Density(dens_arg);
-				dens_arg[0] = PG2;
-				rho_gw = 0.0;
-				if (diffusion)
-				{
-					expfactor =
-					    COMP_MOL_MASS_WATER / (rhow * GAS_CONSTANT * TG);
-					rho_gw = FluidProp->vaporDensity(TG) * exp(-PG * expfactor);
-					p_gw = rho_gw * GAS_CONSTANT * TG / COMP_MOL_MASS_WATER;
-					dens_arg[0] -= p_gw;
-				}
-				rho_ga = GasProp->Density(dens_arg);
-				rho_g = rho_ga + rho_gw;
-			}
-			tensor = MediaProp->PermeabilityTensor(Index);
-			mat_fac = MediaProp->PermeabilitySaturationFunction(Sw, 0) /
-			          FluidProp->Viscosity();
-			m_fac = 0.;
-			if (diffusion)
-				m_fac = rho_gw *
-				        MediaProp->PermeabilitySaturationFunction(Sw, 1) /
-				        (GasProp->Viscosity() * rhow);
-			if (Gravity)
-				mat_fac = mat_fac + m_fac * rho_g / rhow;
-			else
-				mat_fac += m_fac;
-			//
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor *
-				         fac_perm;  // WX:05.2010
-			//
-			if ((!Gravity) && diffusion)
-			{
-				D_gw = tort * COMP_MOL_MASS_WATER * GasProp->molar_mass * M_g *
-				       M_g * rho_g / rhow;
-				D_gw *= time_unit_factor * p_gw / (PG2 * PG2);
-				for (size_t i = 0; i < dim; i++)
-					mat[i * dim + i] -= D_gw;
-			}
-			break;
-		case 2:
-			if (diffusion)
-			{
-				D_ga = tort * COMP_MOL_MASS_WATER * GasProp->molar_mass * M_g *
-				       M_g * rho_g / rhow;
-				D_ga *= time_unit_factor * rho_gw / (PG2 * rhow);
-			}
-			else
-				D_ga = 0.;
-			for (size_t i = 0; i < dim; i++)
-				mat[i * dim + i] = D_ga;
-			break;
-		case 3:
-			// WX: for Cal_Velocity, rho_ga muss be calculated again before
-			// used. 11.05.2010
-			dens_arg[0] = PG2;
-			rho_ga = GasProp->Density(dens_arg);
-			//
-			tensor = MediaProp->PermeabilityTensor(Index);
-			mat_fac = rho_ga *
-			          MediaProp->PermeabilitySaturationFunction(Sw, 1) /
-			          (GasProp->Viscosity() * rhow);
-			//
-			if (Gravity)
-				//        mat_fac *= rhow/rho_ga;
-				mat_fac *= rho_ga / rhow;  // 29.04.2009 WW
-			//
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor *
-				         fac_perm;  // WX:05.2010
-			if ((!Gravity) && diffusion)
-			{
-				D_ga = tort * rho_g * COMP_MOL_MASS_WATER *
-				       GasProp->molar_mass * M_g * M_g / rhow;
-				D_ga *= p_gw / (PG2 * PG2);
-				for (size_t i = 0; i < dim; i++)
-					mat[i * dim + i] += D_ga * time_unit_factor;
-			}
-			break;
-			//------------------------------------------------------------------
-	}
-}
-
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate material coefficient for Laplacian matrix of PS multi-phase
-      flow
-   Programing:
-   03/2009 PCH Implementation
-   last modification:
-**************************************************************************/
-void CFiniteElementStd::CalCoefLaplacePSGLOBAL(bool Gravity, int dof_index)
-{
-	double* tensor = NULL;
-	double mat_fac = 1.0;  // OK411 m_fac=0.;
-	double k_rel = 0.0;
-	double mfp_arg[2];
-	double variables[3];
-
-	int Index = MeshElement->GetIndex();
-	//
-	ComputeShapefct(1);  //  12.3.2007 WW
-	//======================================================================
-	for (size_t i = 0; i < dim * dim; i++)
-		mat[i] = 0.0;
-	switch (dof_index)
-	{
-		case 0:
-			tensor = MediaProp->PermeabilityTensor(Index);
-			if (pcs->m_num->ele_upwinding == 1)
-			{
-				// Doing Upwind elements for saturation by divergent of
-				// pressure.
-				// Pw upwind
-				int WhichNode =
-				    UpwindElement((int)(pcs->m_num->ele_upwind_method), 0);
-				Sw = 1.0 - NodalVal_SatNW[WhichNode];
-			}
-			else
-				Sw = 1.0 - interpolate(NodalVal_SatNW);
-			k_rel = MediaProp->PermeabilitySaturationFunction(Sw, 0);
-
-			// CB_merge_0513
-			variables[0] = interpolate(NodalVal1);  // pressure
-			variables[1] = interpolate(NodalValC);  // temperature
-
-			mat_fac = k_rel / FluidProp->Viscosity(variables);
-			// mat_fac = k_rel / FluidProp->Viscosity();
-			// Since gravity for water phase is handled directly in
-			// Assemble_Gravity,
-			// no need of any code for water phase here.
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor;
-			break;
-		case 1:
-			tensor = MediaProp->PermeabilityTensor(Index);
-			mat_fac = 0.0;  // Snw has no laplace term
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor;
-			break;
-		case 2:
-			tensor = MediaProp->PermeabilityTensor(Index);
-			if (pcs->m_num->ele_upwinding == 1)
-			{
-				// Doing Upwind elements for saturation by divergent of
-				// pressure.
-				// Pnw upwind
-				int WhichNode =
-				    UpwindElement((int)(pcs->m_num->ele_upwind_method), 0);
-				Sw = 1.0 - NodalVal_SatNW[WhichNode];
-			}
-			else
-				Sw = 1.0 - interpolate(NodalVal_SatNW);
-
-			k_rel = MediaProp->PermeabilitySaturationFunction(Sw, 1);
-			// Pnw = Pw + Pc(Sw) //TODO: could cause errors in some cases
-			mfp_arg[0] = interpolate(NodalVal1) +
-			             MediaProp->CapillaryPressureFunction(Sw);
-			mfp_arg[1] = interpolate(NodalValC1);  // TEMPERATURE1 in most cases
-
-			mat_fac = k_rel / GasProp->Viscosity(mfp_arg);
-
-			// The density of the non-wetting phase fluid should be considered
-			// here.
-			// However, the default water phase density should be canceled out
-			// simultaneously.
-			if (Gravity) mat_fac *= GasProp->Density() / FluidProp->Density();
-
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor;
-			break;
-		case 3:
-			// Snw Laplace from Pc in Eqn 2
-			tensor = MediaProp->PermeabilityTensor(Index);
-
-			if (pcs->num_type_name.find("dPcdSwGradSnw") != string::npos)
-			{
-				//			double Snw = -1.0;
-				//			if(pcs->m_num->ele_upwinding == 1)
-				//			{
-				// Doing Upwind elements for saturation by divergent of
-				// pressure.
-				// Pnw upwind
-				//				int WhichNode =
-				// UpwindElement((int)(pcs->m_num->ele_upwind_method), 1); //
-				// TF:
-				// set, but never used
-				//				Snw = NodalVal_SatNW[WhichNode]; // TF: set, but
-				//never
-				// used
-				//			}
-				//			else
-				//				Snw = interpolate(NodalVal_SatNW); // TF: set,
-				//but
-				// never used
-
-				CMediumProperties* m_mmp = NULL;
-				CElem* thisEle = pcs->m_msh->ele_vector[index];
-				int matgrp = thisEle->GetPatchIndex();
-				m_mmp = mmp_vector[matgrp];
-				k_rel = MediaProp->PermeabilitySaturationFunction(Sw, 1);
-
-				double dPcdSw = 0.0;
-				dPcdSw = m_mmp->PressureSaturationDependency(Sw, false);
-
-				// Pnw = Pw + Pc(Sw) // TODO: could cause errors in some cases
-				mfp_arg[0] = interpolate(NodalVal1) +
-				             MediaProp->CapillaryPressureFunction(Sw);
-				mfp_arg[1] = interpolate(NodalValC1);
-				mat_fac = k_rel / GasProp->Viscosity(mfp_arg) * (-dPcdSw);
-			}
-			else
-				mat_fac = 0.0;
-
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor;
-			break;
-		case 4:
-			// For Vnw
-			tensor = MediaProp->PermeabilityTensor(Index);
-
-			//		double Snw = -1.0;
-			//		if(pcs->m_num->ele_upwinding == 1)
-			//		{
-			// Doing Upwind elements for saturation by divergent of pressure.
-			// Pnw upwind
-			//			int WhichNode =
-			// UpwindElement((int)(pcs->m_num->ele_upwind_method), 1); // TF:
-			// set, but never used
-			//			Snw = NodalVal_SatNW[WhichNode]; // TF: set, but never
-			// used
-			//		}
-			//		else
-			//			Snw = interpolate(NodalVal_SatNW); // TF: set, but never
-			// used
-
-			//		CElem* thisEle = pcs->m_msh->ele_vector[index]; // TF: set,
-			// but never used
-			//		int matgrp = thisEle->GetPatchIndex(); // TF: set, but never
-			// used
-			//		CMediumProperties* m_mmp = mmp_vector[matgrp];
-			k_rel = MediaProp->PermeabilitySaturationFunction(Sw, 1);
-			mat_fac = k_rel / GasProp->Viscosity();
-
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac * time_unit_factor;
-			break;
-			//------------------------------------------------------------------
-	}
-}
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   04/2009 PCH Upwind Material Scheme
-   Background:
-   Now material property at the upstream node is taken to exclude future
-   predition at the current due to abrupt change of material properties.
-   This is conservative perticularly the nodes very close to the interface
-   that divides two highly different materials. Thus,this is vector
-   characterisitics. In our case, pressure gradient can be a determinant
-   for permeability and saturation and so other material properties.
-
-   Description:
-   0: none yet
-   1: Upwind element determined by div of pressure
-**************************************************************************/
-int CFiniteElementStd::UpwindElement(int option, int phase)
-{
-	int WhichNodeInTheElement = -1;  // Initialized to be none of elements
-	double Pmin = 1.0 / DBL_MIN;     // Just set to be outrageously big.
-	int GravityOn = 1;               // Initialized to be on
-
-	// If no gravity, then set GravityOn to be zero.
-	if ((coordinate_system) % 10 != 2 && (!axisymmetry)) GravityOn = 0;
-
-	if (option == 1)  // If upwind by divergent of pressure
-	{
-		double PdivAtnode = -1.0 / DBL_MIN;  // Meaningless pressure.
-		double Pc = 0.0;                     // Set to be no capillary at all.
-		int idx_p1 = pcs->GetNodeValueIndex("PRESSURE1");
-		int idx_pc = pcs->GetNodeValueIndex("PRESSURE_CAP");
-
-		for (int i = 0; i < nnodes; ++i)
-		{
-			double Pw = pcs->GetNodeValue(nodes[i], idx_p1 + 1);
-			if (phase == 0)
-			{
-				PdivAtnode = Pw;
-				if (GravityOn)
-					PdivAtnode -= FluidProp->Density() * gravity_constant;
-			}
-			else if (phase == 1)
-			{
-				Pc = pcs->GetNodeValue(nodes[i], idx_pc);
-				PdivAtnode = Pw + Pc;
-				if (GravityOn)
-					PdivAtnode -= GasProp->Density() * gravity_constant;
-			}
-			else
-			{
-				std::cout << "Phase number is wrong in UpwindElement."
-				          << "\n";
-				abort();
-			}
-
-			if (Pmin > PdivAtnode)
-			{
-				Pmin = PdivAtnode;
-				WhichNodeInTheElement = i;
-			}
-		}
-	}
-
-	if (WhichNodeInTheElement == -1)
-	{
-		std::cout << "UpwindElement is failed. Impossible node index!!!"
-		          << "\n";
-		std::cout << "Pmin = " << Pmin << "\n";
-		abort();
-	}
-	return WhichNodeInTheElement;
-}
-// CB 090507
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   05/2007 CB
-   last modification:
-**************************************************************************/
-// void CFiniteElementStd::UpwindUnitCoord(int p, int point, int ind, double
-// *rupw, double *supw, double *tupw)
-void CFiniteElementStd::UpwindUnitCoord(int /*p*/, int point, int ind)
-{
-	double scale;
-	double alpha[3];
-	int gp_r, gp_s, gp_t;
-	bool mmp_element_integration_method_maximum = false;
-
-	int upwind_meth;
-	double upwind_para;
-	double v[3], v_rst[3];
-
-	//
-	ElementValue* gp_ele = ele_gp_value[ind];
-	if (pcs->pcs_type_number == 1)  // WW/CB
-		gp_ele = ele_gp_value[ind + (long)pcs->m_msh->ele_vector.size()];
-
-	// TF unused:  MshElemType::type eletyp = MeshElement->GetElementType();
-
-	//
-	upwind_para = pcs->m_num->ele_upwinding;
-	upwind_meth = pcs->m_num->ele_upwind_method;
-
-	// Numerik
-	// *rupw = *supw = *tupw = 0.0;
-	gp_r = gp_s = gp_t = 0;
-	// alpha initialisieren
-	MNulleVec(alpha, 3);
-	// v initialisieren
-	MNulleVec(v, 3);
-
-	// get the velocities
-	// gp_ele->GetEleVelocity(v);
-
-	// CB: not sure if this is correct, as velocity
-	// at each Gauss point is regarded here (within GP loop)
-	// while in cel_mmp.cpp velocity is evaluated before the GP loop:
-	// CalcVelo3Drst(phase, index, GetTimeCollocationupwind_MMP(), 0., 0., 0.,
-	// v);
-	// v[0] = gp_ele->Velocity(0, point);
-	// v[1] = gp_ele->Velocity(1, point);
-	// v[2] = gp_ele->Velocity(2, point);
-	v[0] = gp_ele->Velocity(0, 0);
-	v[1] = gp_ele->Velocity(1, 0);
-	v[2] = gp_ele->Velocity(2, 0);
-	// this would give v at GP 0
-	// but: if(PcsType==T) SetCenterGP(); // CB 11/2007
-	// was set in Cal_Velo_2(); which is calculated,
-	// when mass matrix is assembled
-	// hence V is at element center of gravity
-	// otherwise use element averaged v?:
-	// for(i=0; i<nGaussPoints; i++)
-	//{
-	//  v[0] += gp_ele->Velocity(0, i)/(double)nGaussPoints;
-	//  v[1] += gp_ele->Velocity(1, i)/(double)nGaussPoints;
-	//  v[2] += gp_ele->Velocity(2, i)/(double)nGaussPoints;
-	//}
-
-	for (size_t i = 0; i < 3; i++)
-		v[i] *= time_unit_factor;
-
-	// instead of r_upw, etc. we use unit[i], next function sets Gauss Integrals
-	// unit[0] = MXPGaussPkt(nGauss, gp_r); -> r_upw ; etc. for unit[i]
-	SetGaussPoint(point, gp_r, gp_s,
-	              gp_t);  // this sets unit[] to standard coordinates
-
-	// v transformation: Jacobi*v-->v_rst
-	// computing the Jacobian at this point results in pressure difference
-	// in comparison to the original model of CT
-	// However, it seems not necessary, as Jacobian has already
-	// been computed in function UpwindAlphaMass
-	// computeJacobian(1); // order 1
-
-	// multiply velocity vector with Jacobian matrix
-	// Jacobi*v-->v_rst
-	// This may need attention to tell different types of elements in the same
-	// dimension	// PCH
-	for (size_t i = 0; i < ele_dim; i++)
-	{
-		v_rst[i] = 0.0;
-		for (size_t j = 0; j < ele_dim; j++)
-			v_rst[i] += Jacobian[i * dim + j] * v[j];
-	}
-	//
-
-	// These need to be rewritten according to different types of elements.  //
-	// PCH
-	if (MBtrgVec(v_rst, ele_dim) > MKleinsteZahl)
-	{
-		// Upwind-Faktoren
-		for (size_t i = 0; i < ele_dim; i++)
-			alpha[i] = -upwind_para * v_rst[i] /
-			           (MBtrgVec(v_rst, ele_dim) + MKleinsteZahl);
-
-		// moving the Gauss points
-		if (upwind_meth == 1)  // limit Gauss point moving on Element domain
-		{
-			scale = 1.;
-			for (size_t i = 0; i < ele_dim; i++)
-			{
-				// Integral over GaussPoints, not used
-				if (mmp_element_integration_method_maximum)
-				{
-					if (fabs(unit[i] + alpha[i]) > 1.)
-						scale =
-						    MMin(scale, (1. - fabs(unit[i])) / fabs(alpha[i]));
-				}
-				else  // regard all quantities in the center of element
-				    if (fabs(alpha[i]) > 1.)
-					scale = MMin(scale, (1. / fabs(alpha[i])));
-			}
-			for (size_t i = 0; i < ele_dim; i++)
-			{
-				// Integral over GaussPoints, not used
-				if (mmp_element_integration_method_maximum)
-					unit[i] +=
-					    scale *
-					    alpha[i];  // scale is added to unit[i] (=Gaussintegral)
-				else  // regard all quantities in the center of element
-					unit[i] = scale * alpha[i];  // unit[i] (=Gaussintegral)
-			}
-		}
-		else if (upwind_meth == 2)  // limit moving on -1<x<1
-
-			// PCH this has never been used, but the code is only for line,
-			// quad, and hex.
-			for (size_t i = 0; i < ele_dim; i++)
-			{
-				// Integral ?er GaussPunkte
-				if (mmp_element_integration_method_maximum)
-					unit[i] = MRange(-1., unit[i] + alpha[i], 1.);
-				else  // regard all quantities in the center of element
-					unit[i] = MRange(-1., alpha[i], 1.);
-			}
-		// here only Methods 1 + 2; M3 = MaxMobilUW is done in CalcCoefLaplace
-	}
-
-#ifdef OLD_UPWINDING
-	// test
-	for (i = 0; i < ele_dim; i++)
-		cout << unit[i] << " ";
-	cout << endl;
-
-	double ur, us, ut;
-
-	switch (eletyp)
-	{
-		case 1:  // Line
-		{
-			// Elementgeometriedaten
-			static double detjac, *invjac, jacobi[4];
-			double l[3];
-			// invjac = GetElementJacobiMatrix(index, &detjac);
-			// Calc1DElementJacobiMatrix(ind, invjac, &detjac);  //ind = element
-			// id number  // wird das irgendwo gebraucht?
-			detjac = computeJacobian(1);  // order
-			invjac = invJacobian;
-
-			MNulleVec(l, 3);
-			l[0] = X[1] - X[0];
-			l[1] = Y[1] - Y[0];
-			l[2] = Z[1] - Z[0];
-
-			// hier nur Methoden 1 + 2; 3 wird in CalcCoefLaplace erledigt
-			if ((upwind_meth == 1) || (upwind_meth == 2))
-			{
-				if (MBtrgVec(v, 3) > MKleinsteZahl)
-				{
-					if (MSkalarprodukt(v, l, 3) > 0.)
-						// CB VZ ge?dert!!!
-						*rupw = MRange(-1., -upwind_para, 1.);
-					//*rupw = MRange(-1., upwind_para , 1.);
-					else
-						// CB VZ ge?dert!!!
-						*rupw = MRange(-1., upwind_para, 1.);
-					//*rupw = MRange(-1., -upwind_para , 1.); //
-				}
-				// else { // test
-				//    cout << "-vau";
-				//    if (MSkalarprodukt(v, l, 3) > 0.)
-				//     *rupw = MRange(-1., upwind_para , 1.);
-				//    else
-				//     *rupw = MRange(-1., -upwind_para , 1.);
-				//}
-				if (aktueller_zeitschritt == 1)  // test
-					*rupw = MRange(-1., upwind_para, 1.);
-			}
-			// Upwind-Faktor Fully upwinding
-		}
-		break;
-		case 2:  // Quadrilateral
-		{
-			// Elementgeometriedaten
-			static double detjac, *invjac, jacobi[4];
-			// Elementdaten
-			static double v_rs[2];
-			// Initialisieren
-			MNulleVec(v_rs, 2);
-
-			// if (mmp_element_integration_method_maximum) ?? CB: was ist das
-			if (1 > 0)
-			{
-				gp_r = (int)(point / nGauss);
-				gp_s = point % nGauss;
-				ur = MXPGaussPkt(nGauss, gp_r);
-				us = MXPGaussPkt(nGauss, gp_s);
-			}
-			else
-			{
-				ur = 0.0;  // Alle Groessen nur in Elementmitte betrachten
-				us = 0.0;  // Alle Groessen nur in Elementmitte betrachten
-			}
-
-			// Geschwindigkeitstransformation: a,b -> r,s
-			// Calc2DElementJacobiMatrix(ind, 0., 0., invjac, &detjac);
-			detjac = computeJacobian(1);  // order
-			invjac = invJacobian;
-			MKopierVec(invjac, jacobi, 4);
-			M2Invertiere(jacobi);  // Jacobi-Matrix
-			MMultMatVec(jacobi, 2, 2, v, 2, v_rs, 2);
-
-			if (MBtrgVec(v_rs, 2) > MKleinsteZahl)
-				// Upwind-Faktoren
-				for (k = 0; k < 2; k++)
-					alpha[k] = -upwind_para * v_rs[k] /
-					           (MBtrgVec(v_rs, 2) + MKleinsteZahl);
-
-			// hier nur Methoden 1 + 2; 3 wird in CalcCoefLaplace erledigt
-			if (upwind_meth == 1)
-			{
-				// Verschiebungen der Gausspunkte auf Element begrenzen
-				scale = 1.;
-				if (fabs(ur + alpha[0]) > 1.)
-					scale = MMin(scale, (1. - fabs(ur)) / fabs(alpha[0]));
-				if (fabs(us + alpha[1]) > 1.)
-					scale = MMin(scale, (1. - fabs(us)) / fabs(alpha[1]));
-				*rupw = ur + scale* alpha[0];
-				*supw = us + scale* alpha[1];
-			}
-			else if (upwind_meth == 2)
-			{
-				// Verschiebungen auf -1<x<1 begrenzen
-				*rupw = MRange(-1., ur + alpha[0], 1.);
-				*supw = MRange(-1., us + alpha[1], 1.);
-			}
-		}
-		break;
-		case 3:  // Hexahedra
-		{
-			/* Elementgeometriedaten */
-			static double detjac, *invjac, jacobi[9];
-			/* Elementdaten */
-			static double v_rst[3];
-			// Initialisieren
-			MNulleVec(v_rst, 3);
-
-			// if (mmp_element_integration_method_maximum) ?? CB: was ist das
-			if (1 > 0)  // CB: to do ??
-			{
-				gp_r = (int)(point / (nGauss * nGauss));
-				gp_s = (point % (nGauss * nGauss));
-				gp_t = gp_s % nGauss;
-				gp_s /= nGauss;
-				ur = MXPGaussPkt(nGauss, gp_r);
-				us = MXPGaussPkt(nGauss, gp_s);
-				ut = MXPGaussPkt(nGauss, gp_t);
-			}
-			else
-			{
-				ur = 0.0;  // Alle Groessen nur in Elementmitte betrachten
-				us = 0.0;  // Alle Groessen nur in Elementmitte betrachten
-				ut = 0.0;  // Alle Groessen nur in Elementmitte betrachten
-			}
-
-			// Calc3DElementJacobiMatrix(ind, 0., 0., 0., invjac, &detjac);
-			detjac = computeJacobian(1);  // order
-			invjac = invJacobian;
-			MKopierVec(invjac, jacobi, 9);
-			M3Invertiere(jacobi); /* zurueck zur Jacobi-Matrix */
-			MMultMatVec(jacobi, 3, 3, v, 3, v_rst, 3);
-
-			if (MBtrgVec(v_rst, 3) > MKleinsteZahl) /* Upwind-Faktoren */
-				for (l = 0; l < 3; l++)
-					alpha[l] = -upwind_para * v_rst[l] / MBtrgVec(v_rst, 3) +
-					           MKleinsteZahl;
-			// hier nur Methoden 1 + 2; 3 wird in CalcCoefLaplace erledigt
-			if (upwind_meth == 1)
-			{
-				// Verschiebungen der Gausspunkte auf Element begrenzen
-				scale = 1.;
-				if (fabs(ur + alpha[0]) > 1.)
-					scale = MMin(scale, (1. - fabs(ur)) / fabs(alpha[0]));
-				if (fabs(us + alpha[1]) > 1.)
-					scale = MMin(scale, (1. - fabs(us)) / fabs(alpha[1]));
-				if (fabs(ut + alpha[2]) > 1.)
-					scale = MMin(scale, (1. - fabs(ut)) / fabs(alpha[2]));
-				*rupw =
-				    ur + scale* alpha[0];  // ist die reihenfolge hier richtig?
-				*supw = us + scale* alpha[1];  // scale h?gt hier ja nur von dem
-				                               // letzten if ab..
-				*tupw = ut + scale* alpha[2];
-			}
-			else if (upwind_meth == 2)
-			{
-				// Verschiebungen auf -1<x<1 begrenzen
-				*rupw = MRange(-1., ur + alpha[0], 1.);
-				*supw = MRange(-1., us + alpha[1], 1.);
-				*tupw = MRange(-1., ut + alpha[2], 1.);
-			}
-		}
-		break;
-		case 4:  // Triangle
-			break;
-		case 5:  // Tedrahedra
-			break;
-		case 6:  // Prism
-			break;
-	}
-
-	// test
-	for (i = 0; i < ele_dim; i++)
-		cout << unit[i] << " ";
-	cout << endl;
-#endif
-}
 
 // SB4200
 /**************************************************************************
@@ -2944,48 +1906,25 @@ void CFiniteElementStd::CalcMass()
 			// Mphi2D_SPG
 			if (pcs->m_num->ele_upwind_method > 0)
 				UpwindSummandMass(gp, gp_r, gp_s, gp_t, alpha, summand);
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//			for (i = 0; i < act_nodes; i++)
-//			{
-//				const int ia = local_idx[i];
-//				for (j = 0; j < nnodes; j++)
-//				{
-//					(*Mass)(ia, j) +=
-//					    mat_fac * (shapefct[ia] + summand[ia]) * shapefct[j];
-//				}
-//			}
-//#else
 			for (i = 0; i < nnodes; i++)
 				for (j = 0; j < nnodes; j++)
-					// bei CT: phi * omega; phi beinh. uw-fakt.
 					(*Mass)(i, j) +=
 					    mat_fac * (shapefct[i] + summand[i]) * shapefct[j];
-//#endif
-			// TEST OUTPUT
 		}
 		else
 #endif
 		{
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//			for (i = 0; i < act_nodes; i++)
-//			{
-//				const int ia = local_idx[i];
-//				for (j = 0; j < nnodes; j++)
-//				{
-//					(*Mass)(ia, j) += mat_fac * shapefct[ia] * shapefct[j];
-//				}
-//			}
-//#else
 			for (i = 0; i < nnodes; i++)
+			{
 				for (j = 0; j < nnodes; j++)
 				{
-					// NW
 					if (pcs->m_num->ele_supg_method == 0)
 						if (j > i) continue;
 					(*Mass)(i, j) += mat_fac * shapefct[i] * shapefct[j];
 				}
-//#endif
-			if (pcs->m_num->ele_supg_method > 0)  // NW
+			}
+
+			if (pcs->m_num->ele_supg_method > 0)
 			{
 				vel[0] = gp_ele->Velocity(0, gp);
 				vel[1] = gp_ele->Velocity(1, gp);
@@ -2994,690 +1933,23 @@ void CFiniteElementStd::CalcMass()
 				double tau = 0;
 				CalcSUPGWeightingFunction(vel, gp, tau, weight_func);
 
-// tau*({v}[dN])^T*[N]
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//				for (i = 0; i < act_nodes; i++)
-//				{
-//					const int ia = local_idx[i];
-//					for (j = 0; j < nnodes; j++)
-//					{
-//						(*Mass)(ia, j) +=
-//						    mat_fac * tau * weight_func[ia] * shapefct[j];
-//					}
-//				}
-//#else
+				// tau*({v}[dN])^T*[N]
 				for (i = 0; i < nnodes; i++)
 					for (j = 0; j < nnodes; j++)
 						(*Mass)(i, j) +=
 						    mat_fac * tau * weight_func[i] * shapefct[j];
-//#endif
 			}
 		}  // end else
 	}      // loop gauss points
 
-// WW/CB //NW
-//#ifndef USE_PETSC
 	if (PcsType != T && pcs->m_num->ele_supg_method == 0)
 	{
 		for (i = 0; i < nnodes; i++)
 			for (j = 0; j < nnodes; j++)
 				if (j > i) (*Mass)(i, j) = (*Mass)(j, i);
 	}
-//#endif
 	// Test Output
 	// Mass->Write();
-}
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   12/2009 NW
-   last modification:
-**************************************************************************/
-double CFiniteElementStd::CalcSUPGCoefficient(double* vel, int ip,
-                                              const double* diff_tensor)
-{
-	//--------------------------------------------------------------------
-	// Collect following information to determine SUPG coefficient
-	// + flow velocity
-	// + diffusivity coefficient (scalar)
-	// + characteristic element length
-	// + (Peclet number)
-
-	double v_mag = sqrt(vel[0] * vel[0] + vel[1] * vel[1] + vel[2] * vel[2]);
-	if (v_mag == 0.0) return 0.0;
-
-	// Characteristic element length
-	double ele_len = CalcSUPGEffectiveElemenetLength(vel);
-	// Diffusivity = (effective heat conductivity) / (fluid heat capacity)
-	const double* dispersion_tensor = diff_tensor;
-	if (diff_tensor == NULL)
-	{
-		if (PcsType == H || PcsType == TH)  // heat
-			dispersion_tensor =
-			    MediaProp->HeatConductivityTensor(MeshElement->GetIndex());
-		// mass
-		else if (PcsType == M)
-			// SB, BG
-			dispersion_tensor = MediaProp->MassDispersionTensorNew(ip, 0);
-	}
-
-	double diff = .0;
-	switch (pcs->m_num->ele_supg_method_diffusivity)
-	{
-		case 1:  // min
-		{
-			double min_diff = dispersion_tensor[0];
-			for (size_t i = 1; i < dim * dim; i++)
-				if (dispersion_tensor[i] < min_diff)
-					min_diff = dispersion_tensor[i];
-			diff = min_diff;
-		}
-		break;
-		case 2:  // magnitude of diagonal
-		{
-			double tmp_diff = 0.0;
-			for (size_t i = 0; i < dim; i++)
-				tmp_diff = dispersion_tensor[i + i * dim] *
-				           dispersion_tensor[i + i * dim];
-			diff = sqrt(tmp_diff);
-		}
-		break;
-		default:  // 0 or any invalid number: max. in dispersion coefficient
-		{
-			double max_diff = dispersion_tensor[0];
-			for (size_t i = 1; i < dim * dim; i++)
-				if (dispersion_tensor[i] > max_diff)
-					max_diff = dispersion_tensor[i];
-			diff = max_diff;
-		}
-		break;
-	}
-
-	if (diff_tensor == NULL)
-	{
-		if (PcsType == H || PcsType == TH)  // heat
-		{
-			if (FluidProp->density_model == 14)
-			{
-				double dens_arg[3];  // AKS
-				int Index = MeshElement->GetIndex();
-				dens_arg[0] = interpolate(NodalValC1);
-				dens_arg[1] = interpolate(NodalVal1) + T_KILVIN_ZERO;
-				dens_arg[2] = Index;
-				diff /= (FluidProp->SpecificHeatCapacity(dens_arg) *
-				         FluidProp->Density(dens_arg));
-			}
-			else
-				diff /=
-				    (FluidProp->SpecificHeatCapacity() * FluidProp->Density());
-		}
-	}
-
-	//--------------------------------------------------------------------
-	// Here calculates SUPG coefficient (tau)
-	double tau = 0.0;
-	switch (pcs->m_num->ele_supg_method)
-	{
-		case 1:
-		{
-			// this coefficient matches with the analytical solution in 1D
-			// steady state case
-			double alpha = 0.5 * v_mag * ele_len / diff;  // 0.5*Pe
-			double func = MLangevin(alpha);
-			tau = 0.5 * ele_len / v_mag * func;
-		}
-		break;
-		case 2:
-		{
-			// taking into account time step
-			//          tau = 1.0 / sqrt(pow(2.0/dt
-			//          ,2.0)+pow(2.0*v_mag/ele_len,2.0));
-			tau = 1.0 / sqrt((2.0 / dt) * (2.0 / dt) +
-			                 (2.0 * v_mag / ele_len) * (2.0 * v_mag / ele_len) +
-			                 (4.0 * diff / (ele_len * ele_len)) *
-			                     (4.0 * diff / (ele_len * ele_len)));
-		}
-		break;
-	}
-
-	return tau;
-}
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   12/2009 NW
-   last modification:
-**************************************************************************/
-void CFiniteElementStd::CalcSUPGWeightingFunction(double* vel, int ip,
-                                                  double& tau, double* v_dN)
-{
-	if (pcs->m_num->ele_supg_method == 0)
-	{
-		cout << "***Warning in CFiniteElementStd::CalcSUPGWeightingFunction(): "
-		        "SUPG option is not selected" << endl;
-		return;
-	}
-
-	// tau
-	tau = CalcSUPGCoefficient(vel, ip);
-
-	// {v}[dN]
-	for (int i = 0; i < nnodes; i++)
-		v_dN[i] = 0.0;
-	for (int i = 0; i < nnodes; i++)
-		for (size_t k = 0; k < dim; k++)
-			v_dN[i] += dshapefct[k * nnodes + i] * vel[k];
-}
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   12/2009 NW
-   last modification:
-**************************************************************************/
-double CFiniteElementStd::CalcSUPGEffectiveElemenetLength(double* /*vel*/)
-{
-	double L = 0.0;
-	switch (this->ele_dim)
-	{
-		case 1:
-		{
-			L = this->MeshElement->GetVolume();
-		}
-		break;
-		case 2:
-		case 3:
-		{
-			switch (pcs->m_num->ele_supg_method_length)
-			{
-				case 1:  // min
-				{
-					double min = MeshElement->GetEdge(0)->getLength();
-					for (size_t i = 1; i < MeshElement->GetEdgesNumber(); i++)
-					{
-						L = MeshElement->GetEdge(i)->getLength();
-						if (L < min) min = L;
-					}
-					L = min;
-				}
-				break;
-				case 2:  // average
-				{
-					double tmp_L = 0.0;
-					for (size_t i = 1; i < MeshElement->GetEdgesNumber(); i++)
-						tmp_L += MeshElement->GetEdge(i)->getLength();
-					L = tmp_L / MeshElement->GetEdgesNumber();
-				}
-				break;
-				case 3:  // stream line length
-				{
-					cout << "***Error: ele_supg_method_length <3> has not been "
-					        "supported yet." << endl;
-				}
-				break;
-				default:  // 0 or any invalid number: max edge length
-				{
-					double max = MeshElement->GetEdge(0)->getLength();
-					for (size_t i = 1; i < MeshElement->GetEdgesNumber(); i++)
-					{
-						L = MeshElement->GetEdge(i)->getLength();
-						if (L > max) max = L;
-					}
-					L = max;
-				}
-				break;
-			}
-		}
-		break;
-	}
-	return L;
-}
-
-// CB 090507
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   05/2007 CB
-   last modification:
-**************************************************************************/
-void CFiniteElementStd::UpwindAlphaMass(double* alpha)
-{
-	// Laufvariablen
-	static long i;
-	// WW int no_phases;
-
-	// static long *element_nodes;
-	// WW double gp[3],
-	double v_rst[3], v_tot[3];
-	// WW static double zeta;
-	// static double *velovec, vg, v[2], vt[2], v_rs[2];
-	// static double alpha_adv[3];
-	double upwind_para;
-	// WW double upwind_meth;
-	// Numerik
-	// WW zeta = 0.0;
-	// WW gp[0]=0.0;   gp[1]=0.0;   gp[2]=0.0;
-
-	int ind = MeshElement->GetIndex();
-	ElementValue* gp_ele = ele_gp_value[ind];
-	//  TF ununsed: MshElemType::type eletyp = MeshElement->GetElementType();
-
-	// Elementdaten und globale Modellparameter
-	// WW no_phases =(int)mfp_vector.size();
-	//
-	upwind_para = pcs->m_num->ele_upwinding;
-	// WW upwind_meth = pcs->m_num->ele_upwind_method;
-
-	// alpha initialisieren
-	MNulleVec(alpha, 3);
-	// v initialisieren
-	MNulleVec(v_tot, 3);
-
-	// get the velocities for phase 0
-	v_tot[0] = gp_ele->Velocity(0, 0);
-	v_tot[1] = gp_ele->Velocity(1, 0);
-	v_tot[2] = gp_ele->Velocity(2, 0);
-	// this would only give v at GP 0
-	// but: if(PcsType==T) SetCenterGP(); // CB 11/2007
-	// was set in Cal_Velo_2(); which is calculated,
-	// when mass matrix is assembled
-	// hence v is at element center of gravity
-	// otherwise use following approximation:
-	// for(i=0; i<nGaussPoints; i++) //element averaged v?
-	//{
-	//  v_tot[0] += gp_ele->Velocity(0, i)/nGaussPoints;
-	//  v_tot[1] += gp_ele->Velocity(1, i)/nGaussPoints;
-	//  v_tot[2] += gp_ele->Velocity(2, i)/nGaussPoints;
-	//}
-
-	// switch to next phase
-	gp_ele = ele_gp_value[ind + (long)pcs->m_msh->ele_vector.size()];
-	// get the velocities for phases 1 and add
-	v_tot[0] += gp_ele->Velocity(0, 0);
-	v_tot[1] += gp_ele->Velocity(1, 0);
-	v_tot[2] += gp_ele->Velocity(2, 0);
-	// for(i=0; i<nGaussPoints; i++) //element averaged v?
-	//{
-	//  v_tot[0] += gp_ele->Velocity(0, i)/nGaussPoints;
-	//  v_tot[1] += gp_ele->Velocity(1, i)/nGaussPoints;
-	//  v_tot[2] += gp_ele->Velocity(2, i)/nGaussPoints;
-	//}
-	for (i = 0; i < 3; i++)
-		v_tot[i] *= time_unit_factor;
-
-	// SetGaussPoint(point, gp_r, gp_s, gp_t);
-	// velocity transformation a,b,c -> r,s,t
-	computeJacobian(1);  // order 1
-	// multiply velocity vector with Jacobian matrix
-	// Jacobi*v-->v_rst
-	for (size_t i = 0; i < ele_dim; i++)
-	{
-		v_rst[i] = 0.0;
-		for (size_t j = 0; j < ele_dim; j++)
-			v_rst[i] += Jacobian[i * dim + j] * v_tot[j];
-	}
-
-	// Upwind-Factors
-	if (MBtrgVec(v_rst, ele_dim) >
-	    MKleinsteZahl)  // if(lengthOftheVector > tolerance)
-
-		for (size_t i = 0; i < ele_dim; i++)
-			alpha[i] = -upwind_para * v_rst[i] /
-			           (MBtrgVec(v_rst, ele_dim) + MKleinsteZahl);
-
-#ifdef OLD_UPWINDING
-
-	// test
-	for (i = 0; i < ele_dim; i++)
-		cout << alpha[i] << " ";
-	cout << endl;
-
-	switch (eletyp)
-	{
-		case 1:  // Line
-		{
-			// Elementgeometriedaten
-			static double detjac, *invjac, jacobi[4];
-			static double l[3];
-			// invjac = GetElementJacobiMatrix(index, &detjac);
-			// Calc1DElementJacobiMatrix(ind, invjac, &detjac);  //index =
-			// element id number
-			detjac = computeJacobian(1);  // order
-			invjac = invJacobian;
-			// element_nodes = ElGetElementNodes(ind);
-
-			MNulleVec(l, 3);
-			l[0] = X[1] - X[0];
-			l[1] = Y[1] - Y[0];
-			l[2] = Z[1] - Z[0];
-
-			if (MBtrgVec(v_tot, 3) > MKleinsteZahl)
-			{
-				if (MSkalarprodukt(v_tot, l, 3) > 0.)
-					zeta = 1.;  // upwind_para
-				else
-					zeta = -1.;  //-upwind_para
-			}
-
-			// aus RF 3.5.06 CT 1D elements: {
-			//// detjac = A*L/2
-			// vorfk = porosity * detjac * Mdrittel;
-			//// Massenmatrix mit SUPG ohne Zeitanteile
-			// mass[0] = (2.0 + 1.5 * mms_upwind_parameter * zeta) * vorfk;
-			// mass[1] = (1.0 + 1.5 * mms_upwind_parameter * zeta) * vorfk;
-			// mass[2] = (1.0 - 1.5 * mms_upwind_parameter * zeta) * vorfk;
-			// mass[3] = (2.0 - 1.5 * mms_upwind_parameter * zeta) * vorfk; // }
-
-			// Upwind-Faktor Fully upwinding
-			// alpha[0]     = m_pcs->m_num->ele_upwinding * zeta;
-			// alpha_adv[0] = m_pcs->m_num->ele_upwinding * zeta;
-			alpha[0] = 1.0 * zeta;  //??
-			                        // alpha_adv[0] = 1.0 * zeta;
-			                        // Advection upwinding
-			// if (MTM2_upwind_method == 2) alpha_adv[0] = ele_upwinding * zeta;
-			// /
-		}
-		break;
-		case 2:  // Quadrilateral
-		{
-			// Elementgeometriedaten
-			static double detjac, *invjac, jacobi[4];
-			// Elementdaten
-			static double v_rs[3];
-
-			// Geschwindigkeitstransformation: a,b -> r,s
-			// Calc2DElementJacobiMatrix(ind, 0., 0., invjac, &detjac);
-			detjac = computeJacobian(1);  // order
-			invjac = invJacobian;
-			MKopierVec(invjac, jacobi, 4);
-			M2Invertiere(jacobi); /* Jacobi-Matrix */
-			MMultMatVec(jacobi, 2, 2, v_tot, 2, v_rs, 2);
-
-			if (MBtrgVec(v_rs, 2) > MKleinsteZahl)
-				// Upwind-Faktoren
-				for (k = 0; k < 2; k++)
-					alpha[k] = -upwind_para * v_rs[k] /
-					           (MBtrgVec(v_rs, 2) + MKleinsteZahl);
-		}
-		break;
-		case 3:  // Hexahedra
-		{
-			/* Elementgeometriedaten */
-			static double* invjac, jacobi[9], detjac;
-			/* Elementdaten */
-			// static double v_rst[3];
-
-			if (MBtrgVec(v_tot, 3) > MKleinsteZahl)
-			{
-				/* Geschwindigkeitstransformation: x,y,z -> r,s,t */
-				// Calc3DElementJacobiMatrix(ind, 0., 0., 0., invjac, &detjac);
-				detjac = computeJacobian(1);  // order
-				invjac = invJacobian;
-				MKopierVec(invjac, jacobi, 9);
-				M3Invertiere(jacobi); /* Jacobi-Matrix */
-				MMultMatVec(jacobi, 3, 3, v_tot, 3, v_rst, 3);
-
-				/* Upwind-Faktoren */
-				for (l = 0; l < 3; l++)
-					alpha[l] = -upwind_para * v_rst[l] /
-					           (MBtrgVec(v_rst, 3) + MKleinsteZahl);
-			}
-		}
-		break;
-		case 4:  // Triangle
-			break;
-		case 5:  // Tedrahedra
-			break;
-		case 6:  // Prism
-			break;
-	}
-	// test
-	for (i = 0; i < ele_dim; i++)
-		cout << alpha[i] << " ";
-	cout << endl;
-#endif
-}
-
-// CB 160507
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   05/2007 CB
-   last modification:
-**************************************************************************/
-void CFiniteElementStd::UpwindSummandMass(const int gp,
-                                          int& gp_r,
-                                          int& gp_s,
-                                          int& gp_t,
-                                          double* alpha,
-                                          double* summand)
-
-{
-	int i;
-	//
-	GetGaussData(gp, gp_r, gp_s, gp_t);  // this sets unit[] to standard values
-	GradShapeFunction(dshapefct, unit);
-	for (i = 0; i < nnodes; i++)
-	{
-		summand[i] = 0.0;
-		for (size_t k = 0; k < dim; k++)
-			summand[i] += dshapefct[nnodes * k + i] * alpha[k];
-		// summand[i] /= (double)nnodes;
-	}
-
-#ifdef OLD_UPWINDING
-
-	double u1, u2, u3;
-	u1 = u2 = u3 = 0;
-
-	MshElemType::type eletyp = MeshElement->GetElementType();
-	switch (eletyp)
-	{
-		case MshElemType::LINE:
-		{
-			// Line
-			gp_r = gp;
-			u1 = MXPGaussPkt(nGauss, gp_r);
-			summand[0] = +alpha[0] * (1 + u1);  // CB: ?? hab ich mir so gedacht
-			summand[1] = -alpha[0] * (1 - u1);  // CB: ?? hab ich mir so gedacht
-			for (i = 0; i < 2; i++)
-				summand[i] *= 0.5;
-		}
-		break;
-		case MshElemType::QUAD:  // Quadrilateral
-		{
-			gp_r = (int)(gp / nGauss);
-			gp_s = gp % nGauss;
-			u1 = MXPGaussPkt(nGauss, gp_r);
-			u2 = MXPGaussPkt(nGauss, gp_s);
-			// derived from MPhi2D_SUPG
-			summand[0] = +alpha[0] * (1 + u2) + alpha[1] * (1 + u1);
-			summand[1] = -alpha[0] * (1 + u2) + alpha[1] * (1 - u1);
-			summand[2] = -alpha[0] * (1 - u2) - alpha[1] * (1 - u1);
-			summand[3] = +alpha[0] * (1 - u2) - alpha[1] * (1 + u1);
-			for (i = 0; i < 4; i++)
-				summand[i] *= 0.25;
-		}
-		break;
-		case MshElemType::HEXAHEDRON:  // Hexahedra
-		{
-			gp_r = (int)(gp / (nGauss * nGauss));
-			gp_s = (gp % (nGauss * nGauss));
-			gp_t = gp_s % nGauss;
-			gp_s /= nGauss;
-			u1 = MXPGaussPkt(nGauss, gp_r);
-			u2 = MXPGaussPkt(nGauss, gp_s);
-			u3 = MXPGaussPkt(nGauss, gp_t);
-			// derived from MPhi3D_SUPG
-			summand[0] = +alpha[0] * (1 + u2) * (1 + u3) +
-			             alpha[1] * (1 + u1) * (1 + u3) +
-			             alpha[2] * (1 + u1) * (1 + u2);
-			summand[1] = -alpha[0] * (1 + u2) * (1 + u3) +
-			             alpha[1] * (1 - u1) * (1 + u3) +
-			             alpha[2] * (1 - u1) * (1 + u2);
-			summand[2] = -alpha[0] * (1 - u2) * (1 + u3) -
-			             alpha[1] * (1 - u1) * (1 + u3) +
-			             alpha[2] * (1 - u1) * (1 - u2);
-			summand[3] = +alpha[0] * (1 - u2) * (1 + u3) -
-			             alpha[1] * (1 + u1) * (1 + u3) +
-			             alpha[2] * (1 + u1) * (1 - u2);
-			summand[4] = +alpha[0] * (1 + u2) * (1 - u3) +
-			             alpha[1] * (1 + u1) * (1 - u3) -
-			             alpha[2] * (1 + u1) * (1 + u2);
-			summand[5] = -alpha[0] * (1 + u2) * (1 - u3) +
-			             alpha[1] * (1 - u1) * (1 - u3) -
-			             alpha[2] * (1 - u1) * (1 + u2);
-			summand[6] = -alpha[0] * (1 - u2) * (1 - u3) -
-			             alpha[1] * (1 - u1) * (1 - u3) -
-			             alpha[2] * (1 - u1) * (1 - u2);
-			summand[7] = +alpha[0] * (1 - u2) * (1 - u3) -
-			             alpha[1] * (1 + u1) * (1 - u3) -
-			             alpha[2] * (1 + u1) * (1 - u2);
-			for (i = 0; i < 8; i++)
-				summand[i] *= 0.125;
-		}
-		break;
-		case MshElemType::TRIANGLE:  // Triangle
-		{
-			// SamplePointTriHQ(gp, unit);
-		}
-		break;
-		case MshElemType::TETRAHEDRON:  // Tedrahedra
-		{
-			// SamplePointTet5(gp, unit);
-		}
-		break;
-		case MshElemType::PRISM:  // Prism
-		{
-			gp_r = gp % nGauss;
-			gp_s = (int)(gp / nGauss);
-			gp_t = (int)(nGaussPoints / nGauss);
-			// u1 = MXPGaussPktTri(nGauss,gp_r,0); //femlib.cpp statt
-			// mathlib.cpp, nicht verfügbar?
-			// u2 = MXPGaussPktTri(nGauss,gp_r,1);
-			// u3 = MXPGaussPkt(gp_t,gp_s);
-		}
-		break;
-	}
-#endif
-}
-
-/***************************************************************************
-   GeoSys - Funktion:
-           CFiniteElementStd:: CalcMass2
-   Programming:
-   02/2007   WW
- **************************************************************************/
-void CFiniteElementStd::CalcMass2()
-{
-	int i, j, in, jn;
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t = 0;
-	double fkt, mat_fac;
-	// Material
-	int dof_n = 2;
-	mat_fac = 1.0;
-	//----------------------------------------------------------------------
-	//======================================================================
-	// Loop over Gauss points
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determinate
-		//---------------------------------------------------------
-		fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		// Compute geometry
-		ComputeShapefct(1);  // Linear interpolation function
-		for (in = 0; in < dof_n; in++)
-			for (jn = 0; jn < dof_n; jn++)
-			{
-				// Material
-				mat_fac = CalCoefMass2(in * dof_n + jn);
-				mat_fac *= fkt;
-				// Calculate mass matrix
-				const int jsh = jn * nnodes;
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//				for (i = 0; i < act_nodes; i++)
-//				{
-//					const int ia = local_idx[i];  // local_idx[ia]
-//					const int ish = ia + in * nnodes;
-//					for (j = 0; j < nnodes; j++)
-//					{
-//						(*Mass2)(ish, j + jsh) +=
-//						    mat_fac * shapefct[ia] * shapefct[j];
-//					}
-//				}
-//#else
-				for (i = 0; i < nnodes; i++)
-				{
-					const int ish = i + in * nnodes;
-					for (j = 0; j < nnodes; j++)
-						(*Mass2)(ish, j + jsh) +=
-						    mat_fac * shapefct[i] * shapefct[j];
-				}
-//#endif
-			}
-	}
-}
-
-/***************************************************************************
-   GeoSys - Funktion:
-           CFiniteElementStd:: CalcMassPSGLOBAL
-   Programming:
-   03/2009   PCH
- **************************************************************************/
-void CFiniteElementStd::CalcMassPSGLOBAL()
-{
-	int i, j, in, jn;
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t = 0;
-	double fkt, mat_fac;
-	// Material
-	int dof_n = 2;
-	mat_fac = 1.0;
-	//----------------------------------------------------------------------
-	//======================================================================
-	// Loop over Gauss points
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determinate
-		//---------------------------------------------------------
-		fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		// Compute geometry
-		ComputeShapefct(1);  // Linear interpolation function
-		for (in = 0; in < dof_n; in++)
-			for (jn = 0; jn < dof_n; jn++)
-			{
-				// Material
-				mat_fac = CalCoefMassPSGLOBAL(in * dof_n + jn);
-				mat_fac *= fkt;
-// Calculate mass matrix
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//				for (i = 0; i < act_nodes; i++)
-//				{
-//					const int ia = local_idx[i];
-//					for (j = 0; j < nnodes; j++)
-//					{
-//						(*Mass2)(ia + in * nnodes, j + jn * nnodes) +=
-//						    mat_fac * shapefct[ia] * shapefct[j];
-//					}
-//				}
-
-//#else
-				for (i = 0; i < nnodes; i++)
-					for (j = 0; j < nnodes; j++)
-						(*Mass2)(i + in * nnodes, j + jn * nnodes) +=
-						    mat_fac * shapefct[i] * shapefct[j];
-//#endif
-			}
-	}
 }
 
 /***************************************************************************
@@ -3723,16 +1995,8 @@ void CFiniteElementStd::CalcLumpedMass()
 	factor = CalCoefMass();
 	pcs->timebuffer = factor;  // Tim Control "Neumann"
 	factor *= vol / (double)nnodes;
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//	for (i = 0; i < act_nodes; i++)
-//	{
-//		const int ia = local_idx[i];
-//		(*Mass)(ia, ia) = factor;
-//	}
-//#else
 	for (i = 0; i < nnodes; i++)
 		(*Mass)(i, i) = factor;
-//#endif
 //
 #ifdef otherLumpedMass
 	int i, j;
@@ -3751,124 +2015,6 @@ void CFiniteElementStd::CalcLumpedMass()
 #endif
 	// TEST OUT
 	// Mass->Write();
-}
-
-///
-/***************************************************************************
-   GeoSys - Funktion:
-           CFiniteElementStd:: CalcLumpedMass2
-   Programming:
-   02/2007   WW Multi-phase flow
- **************************************************************************/
-void CFiniteElementStd::CalcLumpedMass2()
-{
-	int i, in, jn, gp_r, gp_s, gp_t;
-	double factor, vol = 0.0;
-	int dof_n = 2;
-	//----------------------------------------------------------------------
-	// Volume
-	if (axisymmetry)
-	{  // This calculation should be done in CompleteMesh.
-		// However, in order not to destroy the concise of the code,
-		// it is put here. Anyway it is computational cheap. WW
-		vol = 0.0;
-		for (gp = 0; gp < nGaussPoints; gp++)
-			//---------------------------------------------------------
-			//  Get local coordinates and weights
-			//  Compute Jacobian matrix and its determinate
-			//---------------------------------------------------------
-			vol += GetGaussData(gp, gp_r, gp_s, gp_t);
-	}
-	else
-		vol = MeshElement->GetVolume() * MeshElement->GetFluxArea();  // WW. 24.05.2012
-	//----------------------------------------------------------------------
-	// Initialize
-	(*Mass2) = 0.0;
-	// Center of the reference element
-	SetCenterGP();
-	for (in = 0; in < dof_n; in++)
-	{
-		const int ish = in * nnodes;
-		for (jn = 0; jn < dof_n; jn++)
-		{
-			// Factor
-			factor = CalCoefMass2(in * dof_n + jn);
-			pcs->timebuffer = factor;  // Tim Control "Neumann"
-			// Volume
-			factor *= vol / (double)nnodes;
-			const int jsh = jn * nnodes;  // WW
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//			for (i = 0; i < act_nodes; i++)
-//			{
-//				const int ia = local_idx[i];
-//				(*Mass2)(ia + ish, ia + jsh) = factor;
-//			}
-//#else
-			for (i = 0; i < nnodes; i++)
-				(*Mass2)(i + ish, i + jsh) = factor;
-//#endif
-		}
-	}
-	// TEST OUT
-	// Mass2->Write();
-}
-/***************************************************************************
-   GeoSys - Funktion:
-           CFiniteElementStd:: CalcLumpedMassPSGLOBAL
-   Programming:
-   03/2009   PCH PS_GLOBAL for Multi-phase flow
- **************************************************************************/
-void CFiniteElementStd::CalcLumpedMassPSGLOBAL()
-{
-	int i, in, jn, gp_r, gp_s, gp_t;
-	double factor, vol = 0.0;
-	int dof_n = 2;
-	//----------------------------------------------------------------------
-	// Volume
-	if (axisymmetry)
-	{  // This calculation should be done in CompleteMesh.
-		// However, in order not to destroy the concise of the code,
-		// it is put here. Anyway it is computational cheap. WW
-		vol = 0.0;
-		for (gp = 0; gp < nGaussPoints; gp++)
-			//---------------------------------------------------------
-			//  Get local coordinates and weights
-			//  Compute Jacobian matrix and its determinate
-			//---------------------------------------------------------
-			vol += GetGaussData(gp, gp_r, gp_s, gp_t);
-	}
-	else
-		vol = MeshElement->GetVolume();
-	//----------------------------------------------------------------------
-	// Initialize
-	(*Mass2) = 0.0;
-	// Center of the reference element
-	SetCenterGP();
-	for (in = 0; in < dof_n; in++)
-	{
-		const int ish = in * nnodes;  // WW
-		for (jn = 0; jn < dof_n; jn++)
-		{
-			// Factor
-			factor = CalCoefMassPSGLOBAL(in * dof_n + jn);
-			pcs->timebuffer = factor;  // Tim Control "Neumann"
-			// Volume
-			factor *= vol / (double)nnodes;
-			const int jsh = jn * nnodes;  // WW
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//			for (i = 0; i < act_nodes; i++)
-//			{
-//				const int ia = local_idx[i];
-//				(*Mass2)(ia + ish, ia + jsh) = factor;
-//			}
-//#else
-			for (i = 0; i < nnodes; i++)
-				(*Mass2)(i + ish, i + jsh) = factor;
-//#endif
-		}
-	}
-	// TEST OUT
-	//  Mass2->Write();
 }
 
 /***************************************************************************
@@ -3905,21 +2051,9 @@ void CFiniteElementStd::CalcStorage()
 		mat_fac = CalCoefStorage();
 		// GEO factor
 		fkt *= mat_fac;
-// Calculate mass matrix
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//		for (i = 0; i < act_nodes; i++)
-//		{
-//			const int ia = local_idx[i];
-//			for (j = 0; j < nnodes; j++)
-//			{
-//				(*Storage)(ia, j) += fkt * shapefct[ia] * shapefct[j];
-//			}
-//		}
-//#else
 		for (i = 0; i < nnodes; i++)
 			for (j = 0; j < nnodes; j++)
 				(*Storage)(i, j) += fkt * shapefct[i] * shapefct[j];
-//#endif
 	}
 	// TEST OUTPUT
 	//  if(Index == 195){cout << "Storage Matrix: " << "\n"; Storage->Write(); }
@@ -3959,22 +2093,9 @@ void CFiniteElementStd::CalcContent()
 		mat_fac = CalCoefContent();
 		// GEO factor
 		fkt *= mat_fac;
-// Calculate mass matrix
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//		for (i = 0; i < act_nodes; i++)
-//		{
-//			const int ia = local_idx[i];
-//			for (j = 0; j < nnodes; j++)
-//			{
-//				(*Content)(ia, j) += fkt * shapefct[ia] * shapefct[j];
-//			}
-//		}
-
-//#else
 		for (i = 0; i < nnodes; i++)
 			for (j = 0; j < nnodes; j++)
 				(*Content)(i, j) += fkt * shapefct[i] * shapefct[j];
-//#endif
 	}
 }
 
@@ -4049,31 +2170,6 @@ void CFiniteElementStd::CalcLaplace()
 				}
 #endif
 				const int jsh = jn * nnodes;
-//#if defined(USE_PETSC)  // || defined(other parallel libs)//03~04.3012. WW
-//				//---------------------------------------------------------
-//				for (i = 0; i < act_nodes; i++)
-//				{
-//					const int ia = local_idx[i];
-//					const int iish = ia + ish;
-//					for (j = 0; j < nnodes; j++)
-//					{
-//						const int jjsh = j + jsh;
-//						//  if(j>i) continue;
-//						for (k = 0; k < dim; k++)
-//						{
-//							const int ksh = k * nnodes + ia;
-//							const int km = dim * k;
-//							for (l = 0; l < (int)dim; l++)
-//							{
-//								(*Laplace)(iish, jjsh) +=
-//								    fkt * dshapefct[ksh] * mat[km + l] *
-//								    dshapefct[l * nnodes + j];
-//							}
-//						}
-//					}  // j: nodes
-//				}      // i: nodes
-//#else
-				//---------------------------------------------------------
 				for (i = 0; i < nnodes; i++)
 				{
 					const int iish = i + ish;
@@ -4356,25 +2452,13 @@ void CFiniteElementStd::CalcAdvection()
 			             index, m_pcs->GetElementValueIndex("VELOCITY1_Z") + 1);
 		}
 #endif
-//#if defined(USE_PETSC)  //|| defined (other parallel solver)
-//		for (i = 0; i < act_nodes; i++)
-//		{
-//			const int ia = local_idx[i];
-//			for (j = 0; j < nnodes; j++)
-//			{
-//				for (size_t k = 0; k < dim; k++)
-//					(*Advection)(ia, j) +=
-//					    fkt * shapefct[ia] * vel[k] * dshapefct[k * nnodes + j];
-//			}
-//		}
-//#else
 		for (i = 0; i < nnodes; i++)
 			for (j = 0; j < nnodes; j++)
 				for (size_t k = 0; k < dim; k++)
 					(*Advection)(i, j) +=
 					    fkt * shapefct[i] * vel[k] * dshapefct[k * nnodes + j];
-//#endif
-		if (pcs->m_num->ele_supg_method > 0)  // NW
+
+		if (pcs->m_num->ele_supg_method > 0)
 		{
 			vel[0] = gp_ele->Velocity(0, gp);
 			vel[1] = gp_ele->Velocity(1, gp);
@@ -4726,14 +2810,8 @@ void CFiniteElementStd::Assemble_Gravity()
 			}
 			// Calculate mass matrix
 			const int iinn = ii * nnodes;  // 19.06.2012. WW
-//#if defined(USE_PETSC)  //|| defined (other parallel solver) //19.06.2012
-//			for (int ia = 0; ia < act_nodes; ia++)
-//			{
-//				const int i = local_idx[ia];
-//#else
 			for (i = 0; i < nnodes; i++)
 			{
-//#endif
 				const int ipiinn = iinn + i;  // 19.06.2012. WW
 				for (size_t k = 0; k < dim; k++)
 				{
@@ -4772,237 +2850,7 @@ void CFiniteElementStd::Assemble_Gravity()
 	// TEST OUTPUT
 	// RHS->Write();
 }
-/***************************************************************************
-   GeoSys - Funktion:
-           CFiniteElementStd:: Assemby_Gravity
-   Aufgabe:
-           Assemble the contribution of gravity to RHS in Darcy flow
-           to the global system for the pressure equation of multiphase flow
 
-   Programming:
-   10/2008   PCH
- **************************************************************************/
-// Local assembly
-void CFiniteElementStd::Assemble_Gravity_Multiphase()
-{
-	if ((coordinate_system) % 10 != 2 && (!axisymmetry))
-		// 27.2.2007 WW (*GravityMatrix) = 0.0;
-		return;
-	int i, ii;
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t;
-	gp_t = 0;
-	double fkt, rho;  //, rich_f;
-	double k_rel_iteration;
-	// GEO
-	double geo_fac = MediaProp->geo_area;
-	if (!FluidProp->CheckGravityCalculation()) return;
-	long cshift = 0;  // WW
-	//
-	//
-	int dof_n = 1;  // 27.2.2007 WW
-	if (PcsType == V) dof_n = 2;
-
-	// WW 05.01.07
-	cshift = 0;
-	if (pcs->dof > 1) cshift = NodeShift[pcs->continuum];
-
-	// rich_f = 1.0;
-	// if(PcsType==R) rich_f = -1.0; //WW
-
-	k_rel_iteration = 1.0;
-
-	for (i = 0; i < dof_n * nnodes; i++)
-		NodalVal[i] = 0.0;
-
-	// (*GravityMatrix) = 0.0;
-	// Loop over Gauss points
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determination
-		//---------------------------------------------------------
-		fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		//---------------------------------------------------------
-		// Compute geometry
-		//---------------------------------------------------------
-		ComputeGradShapefct(1);  // Linear interpolation function
-		ComputeShapefct(1);      // Moved from CalCoefLaplace(). 12.3.2007 WW
-		// Material
-		// PCH
-		// Pressure equation is the sum of all pressures for all the phases
-		// so is gravity. Thus, this sumation should be considered depending on
-		// solving the pressure or saturation equation.
-		// Thus, the gravity term for the presure equations should cover
-		// two phases.
-		if (pcs->pcs_type_number == 0)  // if the pressure equation,
-		{
-			int numOfPhases = 2;
-			for (int p = 0; p < numOfPhases; ++p)
-			{
-				rho = mfp_vector[p]->Density();
-				if (gravity_constant < MKleinsteZahl)  // HEAD version
-					rho = 1.0;
-				else if (HEAD_Flag)
-					rho = 1.0;
-				else
-					rho *= gravity_constant;
-
-				// Initialization
-				fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-				fkt *= rho;  //*rich_f;
-
-				for (ii = 0; ii < dof_n; ii++)
-				{
-					// CalCoefLaplace does the sumation mat_fac for twophase
-					// flow.
-					// This is not right for two phase gravity terms, because
-					// the equation
-					// is not summable this way. It should be seperated.
-					if (dof_n == 1) CalCoefLaplaceMultiphase(p);
-					if (dof_n == 2) CalCoefLaplace2(false, ii * dof_n + 1);
-
-					// Calculate mass matrix
-					for (i = 0; i < nnodes; i++)
-						for (size_t k = 0; k < dim; k++)
-							NodalVal[i + ii * nnodes] -=
-							    fkt * dshapefct[k * nnodes + i] *
-							    mat[dim * k + dim - 1];
-				}
-			}
-		}
-		else
-		{
-			rho = mfp_vector[1]->Density();
-			if (gravity_constant < MKleinsteZahl)  // HEAD version
-				rho = 1.0;
-			else if (HEAD_Flag)
-				rho = 1.0;
-			else
-				rho *= gravity_constant;
-			fkt *= rho;  //*rich_f;
-
-			for (ii = 0; ii < dof_n; ii++)
-			{
-				// CalCoefLaplace does the sumation mat_fac for twophase flow.
-				// This is not right for two phase gravity terms, because the
-				// equation
-				// is not summable this way. It should be seperated.
-				if (dof_n == 1) CalCoefLaplace(false);
-				if (dof_n == 2) CalCoefLaplace2(false, ii * dof_n + 1);
-
-				// Calculate mass matrix
-				for (i = 0; i < nnodes; i++)
-					for (size_t k = 0; k < dim; k++)
-						NodalVal[i + ii * nnodes] -= fkt *
-						                             dshapefct[k * nnodes + i] *
-						                             mat[dim * k + dim - 1];
-			}
-		}
-	}
-
-	//
-	cshift += NodeShift[problem_dimension_dm];  // 05.01.07 WW
-	int ii_sh = 0;
-	for (ii = 0; ii < dof_n; ii++)  // 07.02.07 WW
-	{
-		cshift += NodeShift[ii];
-		ii_sh = ii * nnodes;
-		for (i = 0; i < nnodes; i++)
-		{
-			eqs_rhs[cshift + eqs_number[i]] +=
-			    k_rel_iteration * geo_fac * NodalVal[i + ii_sh];
-			(*RHS)(i + LocalShift + ii_sh) += NodalVal[i + ii_sh];
-		}
-	}
-	// TEST OUTPUT
-	// RHS->Write();
-}
-////////////////////////////////////////////////////////////////
-/*
-   void  CFiniteElementStd::Assemble_Gravity()
-   {
-   if((coordinate_system)%10!=2)
-     return;
-   int i, j, k, l;
-   // ---- Gauss integral
-   int gp, gp_r=0, gp_s=0, gp_t;
-   gp_t = 0;
-   double fkt, rho;
-   double k_rel_iteration;
-   // GEO
-   double geo_fac = MediaProp->geo_area;
-
-   k_rel_iteration = 1.0;
-
-   (*GravityMatrix) = 0.0;
-   // Loop over Gauss points
-   for (gp = 0; gp < nGaussPoints; gp++)
-   {
-   //---------------------------------------------------------
-   //  Get local coordinates and weights
-   //  Compute Jacobian matrix and its determination
-   //---------------------------------------------------------
-   fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-
-   //---------------------------------------------------------
-   // Compute geometry
-   //---------------------------------------------------------
-   ComputeGradShapefct(1); // Linear interpolation function
-
-   // Material
-   CalCoefLaplace(true);
-   rho = FluidProp->Density(Index,unit,pcs->m_num->time_theta);
-   if(gravity_constant<MKleinsteZahl) // HEAD version
-   rho = 1.0;
-   else if(HEAD_Flag) rho = 1.0;
-   else
-   rho *= gravity_constant;
-
-   fkt *= rho;
-   // Calculate mass matrix
-   for (i = 0; i < nnodes; i++)
-   for (j = 0; j < nnodes; j++)
-   {
-   if(j>i) continue;
-   for (k = 0; k < dim; k++)
-   {
-   for(l=0; l<dim; l++)
-   (*GravityMatrix)(i,j) += fkt*dshapefct[k*nnodes+i]
-   *mat[dim*k+l]* dshapefct[l*nnodes+j];
-   }
-   }
-   }
-
-   //TEST OUTPUT
-   //GravityMatrix->Write();
-
-   double* G_coord = NULL;
-   if((coordinate_system)/10==1)
-   G_coord = X;
-   else if((coordinate_system)/10==2)
-   G_coord = Y;
-   else if((coordinate_system)/10==3)
-   G_coord = Z;
-
-   for (i = 0; i < nnodes; i++)
-   {
-   NodalVal[i] = 0.0;
-   for (j = 0; j < nnodes; j++)
-   NodalVal[i] -= (*GravityMatrix)(i,j)* G_coord[j];
-   }
-
-   for (i=0;i<nnodes;i++)
-   {
-   pcs->eqs->b[NodeShift[problem_dimension_dm] + eqs_number[i]]
-   += k_rel_iteration* geo_fac*NodalVal[i];
-   (*RHS)(i+LocalShift) += NodalVal[i];
-   }
-   //TEST OUTPUT
-   //RHS->Write();
-   }
- */
 
 /***************************************************************************
    GeoSys - Funktion:
@@ -5370,185 +3218,6 @@ double CFiniteElementStd::InterpolatePropertyToGausspoint(int GPIndex,
 
 	return variable;
 }
-
-/***************************************************************************
-      GeoSys - Funktion:
-              CFiniteElementStd:: Velocity calulation
-
-      Programming:  WW
-      08/2005
-      03/2007   WW  Multi-phase flow
-      11/2007   CB  this function was only introduced to allow the calculation
- of
-                    the element center of gravity velocity for upwinding
-
- **************************************************************************/
-// Local assembly
-void CFiniteElementStd::Cal_Velocity_2()
-{
-	int k;
-	static double vel[3], vel_g[3];
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t;
-	double coef = 0.0;
-	int dof_n = 1;
-	if (PcsType == V) dof_n = 2;
-	//
-	gp_t = 0;
-
-	// Get room in the memory for local matrices
-	SetMemory();
-	// Set material
-	SetMaterial();
-
-	ElementValue* gp_ele = ele_gp_value[Index];
-
-	// Loop over Gauss points
-	k = (coordinate_system) % 10;
-	if (PcsType == T)  // WW/CB
-	{
-		if (pcs->pcs_type_number == 0)
-		{
-			// gas pressure
-			idx1 = pcs->GetNodeValueIndex("PRESSURE1") + 1;
-			for (int i = 0; i < nnodes; i++)
-				NodalVal[i] = pcs->GetNodeValue(nodes[i], idx1);
-		}
-		else if (pcs->pcs_type_number == 1)
-		{
-			idxp21 = pcs->GetNodeValueIndex("PRESSURE_CAP");
-			// gas pressure
-			idx1 = cpl_pcs->GetNodeValueIndex("PRESSURE1") + 1;
-			gp_ele = ele_gp_value[Index + (long)pcs->m_msh->ele_vector.size()];
-			for (int i = 0; i < nnodes; i++)
-				// P_l = P_g - P_cap
-				NodalVal[i] = cpl_pcs->GetNodeValue(nodes[i], idx1) -
-				              pcs->GetNodeValue(nodes[i], idxp21);
-		}
-	}
-	else
-		for (int i = 0; i < nnodes; i++)
-			NodalVal[i] = pcs->GetNodeValue(nodes[i], idx1);
-	//
-	if (PcsType == V)
-		for (int i = 0; i < nnodes; i++)
-		{
-			NodalVal[i] -= pcs->GetNodeValue(nodes[i], idxp21);
-			NodalVal1[i] = pcs->GetNodeValue(nodes[i], idxp21);
-		}
-	//
-	gp_ele->Velocity = 0.0;
-	//
-
-	gp = 0;
-
-	// for (gp = 0; gp < nGaussPoints; gp++)
-	//{
-	//---------------------------------------------------------
-	//  Get local coordinates and weights
-	//  Compute Jacobian matrix and its determination
-	//---------------------------------------------------------
-
-	GetGaussData(gp, gp_r, gp_s, gp_t);
-	// calculate the velocity at the element center of gravity
-	if (PcsType == T) SetCenterGP();  // CB 11/2007
-
-	//---------------------------------------------------------
-	// Compute geometry
-	//---------------------------------------------------------
-	ComputeGradShapefct(1);  // Linear interpolation function
-	ComputeShapefct(1);      // Moved from CalCoefLaplace(). 12.3.2007 WW
-	if ((PcsType == T) && (pcs->pcs_type_number == 1))  // WW/CB
-		flag_cpl_pcs = true;
-	// Material
-	if (dof_n == 1)
-		CalCoefLaplace(true);
-	else if (dof_n == 2)
-		CalCoefLaplace2(true, 0);
-	if ((PcsType == T) && (pcs->pcs_type_number == 1))  // WW/CB
-		flag_cpl_pcs = false;
-
-	// Velocity
-	for (size_t i = 0; i < dim; i++)
-	{
-		vel[i] = 0.0;
-		for (int j = 0; j < nnodes; j++)
-			vel[i] += NodalVal[j] * dshapefct[i * nnodes + j];
-		//			 vel[i] += fabs(NodalVal[j])*dshapefct[i*nnodes+j];
-	}
-	if (PcsType == V)
-		for (size_t i = 0; i < dim; i++)
-		{
-			vel_g[i] = 0.0;
-			for (int j = 0; j < nnodes; j++)
-				vel_g[i] += NodalVal1[j] * dshapefct[i * nnodes + j];
-		}
-	// Gravity term
-	if (k == 2 && (!HEAD_Flag))
-	{
-		coef = gravity_constant * FluidProp->Density();
-		if (dim == 3 && ele_dim == 2)
-		{
-			auto &tensor(*MeshElement->getTransformTensor());
-			for (size_t i = 0; i < dim; i++)
-				for (size_t j = 0; j < ele_dim; j++)
-				{
-					vel[i] += coef * tensor(i, k) *
-					          tensor(2, k);
-					if (PcsType == V)
-						vel_g[i] += rho_g * gravity_constant *
-						            tensor(i, k) *
-						            tensor(2, k);
-				}
-		}  // To be correctted
-		else
-		{
-			if (PcsType == V)
-			{
-				vel[dim - 1] -= coef;
-				vel_g[dim - 1] += gravity_constant * rho_g;
-			}
-			else
-				vel[dim - 1] += coef;
-		}
-	}
-	for (size_t i = 0; i < dim; i++)
-		for (size_t j = 0; j < dim; j++)
-			//            gp_ele->Velocity(i, gp) -= mat[dim*i+j]*vel[j];  //
-			//            unit as that given in input file
-			gp_ele->Velocity(i, gp) -=
-			    mat[dim * i + j] * vel[j] / time_unit_factor;
-	//
-	if (PcsType == V)
-	{
-		CalCoefLaplace2(true, 3);
-		coef = rhow / rho_ga;
-		for (size_t i = 0; i < dim; i++)
-			for (size_t j = 0; j < dim; j++)
-				gp_ele->Velocity_g(i, gp) -=
-				    coef * mat[dim * i + j] * vel_g[j] / time_unit_factor;
-	}
-	//
-	//   cout << gp << " " << vel[0] << " " << vel[1] << " " << vel[2] << endl;
-	//   //Test
-	//} // for (gp = 0;...
-	//
-	if (pcs->Write_Matrix)
-	{
-		(*pcs->matrix_file) << "### Element: " << Index << "\n";
-		(*pcs->matrix_file) << "---Velocity of water "
-		                    << "\n";
-		gp_ele->Velocity.Write(*pcs->matrix_file);
-		if (gp_ele->Velocity_g.Size() > 0)
-		{
-			(*pcs->matrix_file) << "---Velocity of gas "
-			                    << "\n";
-			gp_ele->Velocity_g.Write(*pcs->matrix_file);
-		}
-	}
-	// gp_ele->Velocity.Write();
-}
-
 
 /***************************************************************************
    GeoSys - Funktion:
@@ -7889,81 +5558,6 @@ void CFiniteElementStd::CalcNodeMatParatemer()
 	}
 }
 
-// WW 08/2007
-ElementValue::ElementValue(CRFProcess* m_pcs, CElem* ele) : pcs(m_pcs)
-{
-	int NGPoints = 0, NGP = 0;
-	int ele_dim;
-
-	MshElemType::type ele_type = ele->GetElementType();
-	ele_dim = ele->GetDimension();
-
-	NGP = GetNumericsGaussPoints(ele_type);
-	if (ele_type == MshElemType::LINE)
-		// OKWW
-		NGPoints = m_pcs->m_num->ele_gauss_points;
-	else if (ele_type == MshElemType::TRIANGLE)
-		NGPoints = 3;
-	else if (ele_type == MshElemType::TETRAHEDRON)
-		NGPoints = 5;  // 15;
-	else
-		NGPoints = (int)MathLib::fastpow(NGP, ele_dim);
-
-	// WW Velocity.resize(m_pcs->m_msh->GetCoordinateFlag()/10, NGPoints);
-	Velocity.resize(3, NGPoints);
-	Velocity = 0.0;
-	Velocity0.resize(3, NGPoints);
-	Velocity0 = 0.0;
-	// 15.3.2007 Multi-phase flow WW
-	if (pcs->type == 1212 || pcs->type == 1313 || m_pcs->type == 42)
-	{
-		Velocity_g.resize(3, NGPoints);
-		Velocity_g = 0.0;
-	}
-}
-// WW 08/2007
-void ElementValue::getIPvalue_vec(const int IP, double* vec)
-{
-	// SB, BG
-	for (int i = 0; i < int(Velocity.Rows()); i++)
-		vec[i] = Velocity(i, IP);
-}
-// SB, BG 09/2010
-void ElementValue::getIPvalue_vec_phase(const int IP, int phase, double* vec)
-{
-	if (phase == 0)
-		for (int i = 0; (size_t)i < Velocity.Rows(); i++)
-			vec[i] = Velocity(i, IP);
-	else if (phase == 10)
-		for (int i = 0; (size_t)i < Velocity_g.Rows(); i++)
-			vec[i] = Velocity_g(i, IP);
-}
-
-/**************************************************************************
-   FEMLib-Method:
-   Task:
-   Programing:
-   01/2006 YD Implementation
-   last modification:
-**************************************************************************/
-void ElementValue::GetEleVelocity(double* vec)
-{
-	for (int i = 0; (size_t)i < Velocity.Rows(); i++)
-	{
-		vec[i] = 0.0;
-		for (int j = 0; (size_t)j < Velocity.Cols(); j++)
-			vec[i] += Velocity(i, j);
-		vec[i] /= Velocity.Cols();
-	}
-}
-// WW
-ElementValue::~ElementValue()
-{
-	Velocity.resize(0, 0);
-	Velocity0.resize(0, 0);
-	Velocity_g.resize(0, 0);
-}
-
 /**************************************************************************
    FEMLib-Method:
    01/2006 OK Implementation
@@ -8017,216 +5611,6 @@ void CFiniteElementStd::AssembleParabolicEquationRHSVector()
 	}
 #endif
 	//----------------------------------------------------------------------
-}
-
-///////
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate  coefficient of temperature induced RHS of multi-phase
-      flow
-   Programing:
-   02/2007 WW Implementation
-   last modification:
-**************************************************************************/
-double CFiniteElementStd::CalCoef_RHS_T_MPhase(int dof_index)
-{
-	double val = 0.0, D_gw = 0.0, D_ga = 0.0;
-	double expfactor = 0.0, dens_arg[3];
-	int Index = MeshElement->GetIndex();
-	ComputeShapefct(1);
-	//======================================================================
-	switch (dof_index)
-	{
-		case 0:
-			PG = interpolate(NodalVal1);
-			Sw = MediaProp->SaturationCapillaryPressureFunction(PG);
-			TG = interpolate(NodalValC1) + T_KILVIN_ZERO;
-			TG0 = interpolate(NodalValC) + T_KILVIN_ZERO;
-			PG2 = interpolate(NodalVal_p2);
-			rhow = FluidProp->Density();
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			expfactor = COMP_MOL_MASS_WATER / (rhow * GAS_CONSTANT * TG);
-			rho_gw = FluidProp->vaporDensity(TG) * exp(-PG * expfactor);
-			//
-			drho_gw_dT = (FluidProp->vaporDensity_derivative(TG) +
-			              PG * expfactor * FluidProp->vaporDensity(TG) / TG) *
-			             exp(-PG * expfactor);
-			val = (1. - Sw) * poro * drho_gw_dT / rhow;
-			//
-			if (SolidProp)
-				val -= (1.0 - poro) * ((1 - Sw) * rho_gw / rhow + Sw) *
-				       SolidProp->Thermal_Expansion();
-			//
-			// val += n*(1.0-rho_gw/rhow)*(dSw/dT)
-			val *= (TG - TG0);
-			break;
-		case 1:
-			//
-			val = -(1. - Sw) * poro * drho_gw_dT / rhow;
-			//
-			if (SolidProp)
-				val -= (1.0 - poro) * (1 - Sw) * rho_ga *
-				       SolidProp->Thermal_Expansion() / rhow;
-			//
-			// val -= n*rho_ga/rhow)*(dSw/dT)
-			//---------------------------------------------------------------
-			val *= (TG - TG0);
-			break;
-		case 2:
-			//------------------------------------------------------------------------
-			// From grad (p_gw/p_g)
-			tort = MediaProp->TortuosityFunction(Index, unit,
-			                                     pcs->m_num->time_theta);
-			tort *= MediaProp->base_heat_diffusion_coefficient * (1 - Sw) *
-			        poro * pow(TG / T_KILVIN_ZERO, 1.8);
-			p_gw = rho_gw * GAS_CONSTANT * TG / COMP_MOL_MASS_WATER;
-			dens_arg[0] = PG2 - p_gw;
-			dens_arg[1] = TG;
-			rho_ga = GasProp->Density(dens_arg);  // AKS SEP 2010
-			// //(PG2-p_gw)*GasProp->molar_mass/(GAS_CONSTANT*TG);
-			rho_g = rho_ga + rho_gw;
-			// 1/Mg
-			M_g =
-			    (rho_gw / COMP_MOL_MASS_WATER + rho_ga / GasProp->molar_mass) /
-			    rho_g;
-			D_gw = tort * rho_g * COMP_MOL_MASS_WATER * GasProp->molar_mass *
-			       M_g * M_g / rhow;
-			val = D_gw * drho_gw_dT * GAS_CONSTANT * TG /
-			      (COMP_MOL_MASS_WATER * PG2) * time_unit_factor;
-			break;
-		case 3:
-			//---------------------------------------------------------------
-			//
-			D_ga = tort * rho_g * COMP_MOL_MASS_WATER * GasProp->molar_mass *
-			       M_g * M_g / rhow;
-			// From grad (p_gw/p_g)
-			val = -D_ga * drho_gw_dT * GAS_CONSTANT * TG /
-			      (COMP_MOL_MASS_WATER * PG2) * time_unit_factor;
-
-			break;
-			//------------------------------------------------------------------
-	}
-	return val;
-}
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate coefficient of temperature induced RHS of PSGlobal scheme
-
-   Programing:
-
-   last modification:
-**************************************************************************/
-double CFiniteElementStd::CalCoef_RHS_T_PSGlobal(int dof_index)
-{
-	double val = 0.0, variables[3];  // OK411 D_gw=0.0, D_ga=0.0;
-	// OK411 double expfactor=0.0;
-	double P, T;
-	int Index = MeshElement->GetIndex();
-	ComputeShapefct(1);
-	//======================================================================
-	switch (dof_index)
-	{
-		case 0:
-			val = 0.;
-			break;
-		case 1:
-			poro = MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			Sw = 1.0 - interpolate(NodalVal_SatNW);
-			// Pnw = Pw + Pc(Sw)
-			P = interpolate(NodalVal1) +
-			    MediaProp->CapillaryPressureFunction(Sw);
-			//      P  = interpolate(NodalVal1);  // Pw
-			T = interpolate(NodalValC1);
-			variables[0] = P;
-			variables[1] = T;
-			val = -(1. - Sw) * poro * GasProp->drhodT(variables) /
-			      GasProp->Density();
-			break;
-		case 2:
-			val = 0.;
-			break;
-		case 3:
-			val = 0.;
-			break;
-			//------------------------------------------------------------------
-	}
-	return val;
-}
-
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate  Capillary pressure for RHS in the global scheme
-   Programing:
-   03/2009 PCH Implementation
-   last modification:
-**************************************************************************/
-void CFiniteElementStd::CalCoef_RHS_Pc(int dof_index)
-{
-	double* tensor = NULL;
-	double mat_fac = 1.0;  // OK411 m_fac=0.;
-	double k_rel = 0.0;
-
-	int Index = MeshElement->GetIndex();
-	//
-	ComputeShapefct(1);  //  12.3.2007 WW
-	//======================================================================
-	for (size_t i = 0; i < dim * dim; i++)
-		mat[i] = 0.0;
-
-	switch (dof_index)
-	{
-		case 0:
-			break;
-		case 1:
-			break;
-		case 2:
-			tensor = MediaProp->PermeabilityTensor(Index);
-			Sw = 1.0 - interpolate(NodalVal_SatNW);
-			k_rel = MediaProp->PermeabilitySaturationFunction(Sw, 1);
-			mat_fac = k_rel / GasProp->Viscosity() * time_unit_factor;
-
-			for (size_t i = 0; i < dim * dim; i++)
-				mat[i] = tensor[i] * mat_fac;
-			break;
-		case 3:
-			break;
-	}
-}
-
-/**************************************************************************
-   FEMLib-Method:
-   Task: Calculate coefficient of Pc induced RHS of multi-phase
-      flow
-   Programing:
-   03/2009 PCH Implementation
-   last modification:
-**************************************************************************/
-double CFiniteElementStd::CalCoef_RHS_PSGLOBAL(int dof_index)
-{
-	ComputeShapefct(1);
-	for (size_t i = 0; i < dim * dim; i++)
-		mat[i] = 0.0;
-
-	switch (dof_index)
-	{
-		case 0:
-			return 0.0;
-			break;
-		case 1:
-			return 0.0;
-			break;
-		case 2:
-		{
-			Sw = 1.0 - interpolate(NodalVal_SatNW);
-			double k_rel(MediaProp->PermeabilitySaturationFunction(Sw, 1));
-			return (k_rel / GasProp->Viscosity() * time_unit_factor);
-			break;
-		}
-		case 3:
-			return 0.0;
-			break;
-	}
-	return 0.0;
 }
 
 /**************************************************************************
@@ -8293,233 +5677,7 @@ double CFiniteElementStd::CalCoef_RHS_HEAT_TRANSPORT(int dof_index)
 	}
 	return val;
 }
-/***************************************************************************
-   GeoSys - Funktion:
-          Assemble_RHS_T_MPhaseFlow
-   Programming:
-   02/2007   WW
- **************************************************************************/
-void CFiniteElementStd::Assemble_RHS_T_MPhaseFlow()
-{
-	int i, j, ii;
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t = 0;
-	double fkt, fac;
-	// Material
-	int dof_n = 2;
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-	// 02.2011 WW
-	int dm_shift = 0;
-	if (pcs->type / 10 == 4) dm_shift = problem_dimension_dm;
-#endif
-	//----------------------------------------------------------------------
-	for (i = 0; i < dof_n * nnodes; i++)
-		NodalVal[i] = 0.0;
-	//======================================================================
-	// Loop over Gauss points
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determinate
-		//---------------------------------------------------------
-		fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		// Compute geometry
-		ComputeGradShapefct(1);  // Linear interpolation function
-		ComputeShapefct(1);      // Linear interpolation function
-		for (ii = 0; ii < dof_n; ii++)
-		{
-			// Material
-			fac = fkt * CalCoef_RHS_T_MPhase(ii) / dt;
-			// Calculate THS
-			for (i = 0; i < nnodes; i++)
-				NodalVal[i + ii * nnodes] += fac * shapefct[i];
-		}
-		// grad T
-		for (ii = 0; ii < dof_n; ii++)
-		{
-			// Material
-			fac = fkt * CalCoef_RHS_T_MPhase(ii + dof_n);
-			// Calculate THS
-			for (i = 0; i < nnodes; i++)
-				for (j = 0; j < nnodes; j++)
-					for (size_t k = 0; k < dim; k++)
-						NodalVal[i + ii * nnodes] +=
-						    fac * dshapefct[k * nnodes + i] *
-						    dshapefct[k * nnodes + j] *
-						    (NodalValC1[j] + T_KILVIN_ZERO);
-		}
-	}
-	int ii_sh;
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-	long i_sh;
-#endif
-	for (ii = 0; ii < 2; ii++)
-	{
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-		i_sh = NodeShift[ii + dm_shift];
-#endif
-		ii_sh = ii * nnodes;
-		for (i = 0; i < nnodes; i++)
-		{
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-			eqs_rhs[i_sh + eqs_number[i]] -= NodalVal[i + ii_sh];
-#endif
-			(*RHS)(i + LocalShift + ii_sh) -= NodalVal[i + ii_sh];
-		}
-	}
-	//
-}
-/***************************************************************************
-   GeoSys - Function: Assemble_RHS_T_PSGlobal
-   Programming:
-   09/2009
- **************************************************************************/
-void CFiniteElementStd::Assemble_RHS_T_PSGlobal()
-{
-	int i, ii;  // OK411 j, k,
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t = 0;
-	double fkt, fac;
-	// Material
-	int dof_n = 2;
-	double temp[10];
 
-	for (i = 0; i < 10; i++)
-		temp[i] = 0;  // remove
-
-	//----------------------------------------------------------------------
-	for (i = 0; i < dof_n * nnodes; i++)
-		NodalVal[i] = 0.0;
-	//======================================================================
-	// Loop over Gauss points
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determinate
-		//---------------------------------------------------------
-		fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		// Compute geometry
-		ComputeGradShapefct(1);  // Linear interpolation function
-		ComputeShapefct(1);      // Linear interpolation function
-
-		for (ii = 0; ii < dof_n; ii++)
-		{
-			// Material
-			fac = fkt * CalCoef_RHS_T_PSGlobal(ii) / dt;
-			// Calculate THS
-			for (i = 0; i < nnodes; i++)
-			{
-				NodalVal[i + ii * nnodes] += fac * shapefct[i];
-				// remove
-				temp[i + ii * nnodes] += fac * shapefct[i];
-			}
-		}
-	}
-	int ii_sh;
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-	long i_sh;
-#endif
-	for (ii = 0; ii < pcs->dof; ii++)
-	{
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-		i_sh = NodeShift[ii];
-#endif
-		ii_sh = ii * nnodes;
-		for (i = 0; i < nnodes; i++)
-		{
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-			eqs_rhs[i_sh + eqs_number[i]] -= NodalVal[i + ii_sh];
-#endif
-			(*RHS)(i + LocalShift + ii_sh) -= NodalVal[i + ii_sh];
-		}
-	}
-	//
-}
-
-/***************************************************************************
-   GeoSys - Funktion:
-          Assemble_RHS_Pc
-   Programming:
-   03/2009   PCH
- **************************************************************************/
-void CFiniteElementStd::Assemble_RHS_Pc()
-{
-	int i, j, ii;
-	// ---- Gauss integral
-	int gp_r = 0, gp_s = 0, gp_t = 0;
-	double fkt;
-	// Material
-	int dof_n = 2;
-	int ndx_p_cap = pcs->GetNodeValueIndex("PRESSURE_CAP");
-	//----------------------------------------------------------------------
-
-	//      double temp[20];
-
-	for (i = 0; i < dof_n * nnodes; i++)
-	{
-		//         temp[i] = NodalVal[i] = 0.0;
-		NodalVal[i] = 0.0;
-		NodalVal1[i] = 0.0;
-	}
-	for (i = 0; i < nnodes; i++)
-		//        temp[i+dof_n] = NodalVal1[i+dof_n] =
-		//        -pcs->GetNodeValue(nodes[i],ndx_p_cap);
-		NodalVal1[i + dof_n] = -pcs->GetNodeValue(nodes[i], ndx_p_cap);
-
-	//======================================================================
-	// Loop over Gauss points
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determinate
-		//---------------------------------------------------------
-		fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		// Compute geometry
-		ComputeGradShapefct(1);  // Linear interpolation function
-		ComputeShapefct(1);      // Linear interpolation function
-		// grad Pc
-		for (ii = 0; ii < dof_n; ii++)
-		{
-			// Material
-			CalCoef_RHS_Pc(ii + dof_n);
-			// Calculate Pc
-			for (i = 0; i < nnodes; i++)
-				for (j = 0; j < nnodes; j++)
-					for (size_t k = 0; k < dim; k++)
-						for (size_t l = 0; l < dim; l++)
-							NodalVal[dof_n + i] += fkt * mat[dim * k + l] *
-							                       dshapefct[k * nnodes + i] *
-							                       dshapefct[l * nnodes + j] *
-							                       NodalVal1[j + dof_n];
-		}
-	}
-
-	//      for(i=0; i<2*nnodes; ++i)
-	//         temp[i]=NodalVal[i];
-
-	int ii_sh;
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-	long i_sh;
-#endif
-	for (ii = 0; ii < pcs->dof; ii++)
-	{
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-		i_sh = NodeShift[ii];
-#endif
-		ii_sh = ii * nnodes;
-		for (i = 0; i < nnodes; i++)
-		{
-#if !defined(USE_PETSC)  // && !defined(other parallel libs)//03~04.3012. WW
-			eqs_rhs[i_sh + eqs_number[i]] += NodalVal[i + ii_sh];
-#endif
-			(*RHS)(i + LocalShift + ii_sh) += NodalVal[i + ii_sh];
-		}
-	}
-	//
-}
 
 /***************************************************************************
    GeoSys - Funktion:
@@ -9098,62 +6256,6 @@ void CFiniteElementStd::Assemble_RHS_HEAT_TRANSPORT2()
 }
 
 /**************************************************************************
-   FEMLib-Method:
-   Task: Calculate  coefficient of displacement induced RHS of multi-phase
-      flow
-   Programing:
-   02/2007 WW Implementation
-   05/2008 WW Generalization
-   last modification:
-**************************************************************************/
-double CFiniteElementStd::CalCoef_RHS_M_MPhase(int dof_index)
-{
-	double val = 0.0;
-	double expfactor = 0.0;
-	double dens_aug[3];
-	dens_aug[1] = 293.15;
-	bool diffusion = false;  // 08.05.2008 WW
-	if (MediaProp->heat_diffusion_model == 1 && cpl_pcs) diffusion = true;
-	//======================================================================
-	switch (dof_index)
-	{
-		case 0:
-			PG = interpolate(NodalVal1);
-			PG2 = interpolate(NodalVal_p2);  // JT
-			Sw = MediaProp->SaturationCapillaryPressureFunction(PG);
-			if (diffusion)
-			{
-				TG = interpolate(NodalValC1) + T_KILVIN_ZERO;
-				dens_aug[1] = TG;
-			}
-			//
-			dens_aug[0] =
-			    PG2 -
-			    PG;  // JT: this was wrong. density argument is water pressure.
-			rhow = FluidProp->Density(dens_aug);
-			val = Sw;
-			//
-			dens_aug[0] = PG2;
-			//
-			if (diffusion)
-			{
-				expfactor = COMP_MOL_MASS_WATER / (rhow * GAS_CONSTANT * TG);
-				rho_gw = FluidProp->vaporDensity(TG) * exp(-PG * expfactor);
-				p_gw = rho_gw * GAS_CONSTANT * TG / COMP_MOL_MASS_WATER;
-				dens_aug[0] -= p_gw;
-			}
-			rho_ga = GasProp->Density(dens_aug);
-			if (diffusion) val += (1.0 - Sw) * rho_gw / rhow;
-			break;
-		case 1:
-			val = (1.0 - Sw) * rho_ga / rhow;
-			break;
-			//------------------------------------------------------------------
-	}
-	return val;
-}
-
-/**************************************************************************
    PCSLib-Method:
    01/2007 OK Implementation
 **************************************************************************/
@@ -9709,1216 +6811,5 @@ void CFiniteElementStd::PrintTheSetOfElementMatrices(std::string mark)
 	}
 }
 
-void CFiniteElementStd::AssembleTHEquation(bool updateA, bool updateRHS)
-{
-	const unsigned c_dim = dim;
-	const int c_nnodes = nnodes;
-	const double pcs_time_step = pcs->Tim->time_step_length;
-	const double dt_inverse = 1.0 / pcs_time_step;
-	const double theta = pcs->m_num->time_theta;
-	const bool hasGravity =
-	    (coordinate_system) % 10 == 2 && FluidProp->CheckGravityCalculation();
-	const double g_const = hasGravity ? gravity_constant : .0;
-	const bool useSUPG = (pcs->m_num->ele_supg_method > 0);
-	const bool useLumpedMass = (pcs->m_num->ele_mass_lumping > 0);
-	const bool isTransient =
-	    (this->pcs->tim_type == FiniteElement::TIM_TRANSIENT);
-	//	ElementValue* gp_ele = ele_gp_value[Index];
-	static Matrix t_transform_tensor(3, 3);
-	if (dim > MediaProp->geo_dimension)
-	{
-		if (MeshElement->getTransformTensor() == NULL)
-		{
-			std::cout << "***Error: Geometric dimension in MMP is not "
-			             "consistent with element."
-			          << "\n";
-			exit(0);
-		}
-		t_transform_tensor.LimitSize(dim, dim);
-		MeshElement->getTransformTensor()->GetTranspose(t_transform_tensor);
-	}
-
-	for (int i = 0; i < nnodes; i++)
-	{
-		NodalVal_p0[i] = pcs->GetNodeValue(nodes[i], idxp0);
-		NodalVal_p1[i] = pcs->GetNodeValue(nodes[i], idxp1);
-		NodalVal_T0[i] = pcs->GetNodeValue(nodes[i], idxT0);
-		NodalVal_T1[i] = pcs->GetNodeValue(nodes[i], idxT1);
-	}
-
-	// Calculate matrices
-	const int offset_p = 0;
-	const int offset_T = c_nnodes;
-	(*StiffMatrix) = 0.0;  // Jacobian
-	(*RHS) = 0.0;          // Residual
-                           //	if (myrank==2 && updateA) {
-                           //		StiffMatrix->Write();
-                           //	}
-
-//#define TH_DEBUG
-#ifdef TH_DEBUG
-	// for debugging
-	Mass->LimitSize(nnodes, nnodes);
-	Laplace->LimitSize(nnodes, nnodes);
-	Advection->LimitSize(nnodes, nnodes);
-	(*Mass) = .0;
-	(*Laplace) = .0;
-	(*Advection) = .0;
-#endif
-#define ELE_CONST_K
-#ifdef ELE_CONST_K
-	double* tmp_k_tensor = MediaProp->PermeabilityTensor(Index);
-	if (c_dim > MediaProp->geo_dimension)
-	{
-		Matrix local_tensor(c_dim, c_dim), temp_tensor(c_dim, c_dim),
-		    global_tensor(c_dim, c_dim);
-		const unsigned c_ele_dim = ele_dim;
-		for (size_t i = 0; i < c_ele_dim; i++)
-			for (size_t j = 0; j < c_ele_dim; j++)
-				local_tensor(i, j) = tmp_k_tensor[j + i * c_ele_dim];
-		// cout << "K':" << endl; local_tensor.Write();
-		local_tensor.multi(t_transform_tensor, temp_tensor);
-		for (size_t i = 0; i < c_dim; i++)
-			for (size_t j = 0; j < c_dim; j++)
-				for (size_t k = 0; k < c_dim; k++)
-					global_tensor(i, j) +=
-					    (*MeshElement->getTransformTensor())(i, k) *
-					    temp_tensor(k, j);
-		// cout << "K:" << endl; global_tensor.Write();
-		for (size_t i = 0; i < c_dim; i++)
-			for (size_t j = 0; j < c_dim; j++)
-				tmp_k_tensor[c_dim * i + j] = global_tensor(i, j);
-	}
-	double const* const k_tensor = tmp_k_tensor;
-	const double Ss =
-	    MediaProp->StorageFunction(Index, unit, pcs->m_num->time_theta);
-	double dummy[3] = {};
-	double const* const lambda_tensor =
-	    MediaProp->HeatDispersionTensorNew(0, dummy);
-#endif
-
-#define USE_LMASS1
-#ifdef USE_LMASS1
-	if (isTransient && useLumpedMass)
-	{
-		const double vol =
-		    MeshElement->GetVolume() * MeshElement->GetFluxArea();
-		// Center of the reference element
-		SetCenterGP();
-		ComputeShapefct(1);
-		const double gp_p1 = interpolate(NodalVal_p1);
-		const double gp_T1 = interpolate(NodalVal_T1);
-		double var[3] = {};
-		var[0] = gp_p1;
-		var[1] = gp_T1;
-		const double rhocp =
-		    MediaProp->HeatCapacity(Index, pcs->m_num->time_theta, this, var);
-		const double drho_w_dp = FluidProp->drhodP(var);
-		const double drho_w_dT = FluidProp->drhodT(var);
-		const double cp_w = FluidProp->SpecificHeatCapacity(var);
-		const double porosity =
-		    MediaProp->Porosity(Index, pcs->m_num->time_theta);
-		const double drhocp_dp = porosity * cp_w * drho_w_dp;
-		const double drhocp_dT = porosity * cp_w * drho_w_dT;
-		const double fkt = vol / (double)nnodes;
-		const double M_tt_coeff1 = fkt * dt_inverse * rhocp;
-		const double M_tt_coeff2 = fkt * dt_inverse * drhocp_dT;
-		const double M_tp_coeff2 = fkt * dt_inverse * drhocp_dp;
-		if (updateA)
-			for (int i = 0; i < nnodes; i++)
-			{
-				(*StiffMatrix)(offset_T + i, offset_T + i) += M_tt_coeff1;
-				for (int j = 0; j < nnodes; j++)
-				{
-					(*StiffMatrix)(offset_T + i, offset_T + j) +=
-					    M_tt_coeff2 * (NodalVal_T1[i] - NodalVal_T0[i]) *
-					    shapefct[j];
-					(*StiffMatrix)(offset_T + i, offset_p + j) +=
-					    M_tp_coeff2 * (NodalVal_T1[i] - NodalVal_T0[i]) *
-					    shapefct[j];
-				}
-			}
-		if (updateRHS)
-			for (int i = 0; i < nnodes; i++)
-			{
-				(*RHS)(offset_T + i) +=
-				    M_tt_coeff1 * (NodalVal_T1[i] - NodalVal_T0[i]);
-			}
-	}
-#endif
-
-	//======================================================================
-	// Loop over Gauss points
-	int gp_r, gp_s, gp_t;
-	double var[3] = {}, vel[3] = {};
-	for (gp = 0; gp < nGaussPoints; gp++)
-	{
-		//---------------------------------------------------------
-		//  Get local coordinates and weights
-		//  Compute Jacobian matrix and its determinate
-		//---------------------------------------------------------
-		const double fkt = GetGaussData(gp, gp_r, gp_s, gp_t);
-		// Compute geometry
-		ComputeShapefct(1);
-		ComputeGradShapefct(1);
-		double const* const c_shapefct = shapefct;
-		double const* const c_dshapefct = dshapefct;
-
-		//---------------------------------------------------------
-		//  Get state variables
-		//---------------------------------------------------------
-		const double gp_p0 = interpolate(NodalVal_p0);
-		const double gp_T0 = interpolate(NodalVal_T0);
-		const double gp_p1 = interpolate(NodalVal_p1);
-		const double gp_T1 = interpolate(NodalVal_T1);
-		//		if (myrank==2 && gp==0 && updateA)
-		//		ScreenMessage2("%d: p0=%.3e, p1=%.3e, T0=%.3e, T1=%.3e\n",
-		//Index,
-		// gp_p0, gp_p1, gp_T0, gp_T1);
-		var[0] = gp_p1;
-		var[1] = gp_T1;
-		double grad_p0[3] = {}, grad_T0[3] = {};
-		double grad_p1[3] = {}, grad_T1[3] = {};
-		double grad_T1_r[3] = {};
-		//		static double max_gradT1 = .0;
-		for (unsigned i = 0; i < c_dim; i++)
-		{
-			for (int j = 0; j < c_nnodes; j++)
-			{
-				grad_p0[i] += c_dshapefct[i * nnodes + j] * NodalVal_p0[j];
-				grad_p1[i] += c_dshapefct[i * nnodes + j] * NodalVal_p1[j];
-				grad_T0[i] += c_dshapefct[i * nnodes + j] * NodalVal_T0[j];
-				grad_T1_r[i] += c_dshapefct[i * nnodes + j] * NodalVal_T1[j];
-			}
-#define TH_INCLUDE_GRAD_T
-#ifdef TH_INCLUDE_GRAD_T
-			if (pcs->m_num->nls_jacobian_level == 0) grad_T1[i] = grad_T1_r[i];
-#ifdef USE_PETSC
-//			if (fabs(grad_T1[i]) > fabs(max_gradT1) ) {
-//				max_gradT1 = grad_T1[i];
-////				ScreenMessage2("-> max grad T updated = %g\n", max_gradT1);
-//			}
-#endif
-#endif
-		}
-
-//---------------------------------------------------------
-//  Get material properties
-//---------------------------------------------------------
-#ifdef CONST_HEAT
-		{
-#endif
-			// Fluid properties
-			const double rho_w = FluidProp->Density(var);
-			const double drho_w_dp = FluidProp->drhodP(var);
-			const double drho_w_dT = FluidProp->drhodT(var);
-			const double vis = FluidProp->Viscosity(var);
-			const double dvis_dp = FluidProp->dViscositydP(var);
-			const double dvis_dT = FluidProp->dViscositydT(var);
-			const double cp_w = FluidProp->SpecificHeatCapacity(var);
-// Medium properties
-#ifndef ELE_CONST_K
-			const double Ss =
-			    MediaProp->StorageFunction(Index, unit, pcs->m_num->time_theta);
-			double* k_tensor = MediaProp->PermeabilityTensor(Index);
-			if (dim > MediaProp->geo_dimension)
-			{
-				Matrix local_tensor(dim, dim), temp_tensor(dim, dim),
-				    global_tensor(dim, dim);
-				for (size_t i = 0; i < ele_dim; i++)
-					for (size_t j = 0; j < ele_dim; j++)
-						local_tensor(i, j) = k_tensor[j + i * ele_dim];
-				// cout << "K':" << endl; local_tensor.Write();
-				local_tensor.multi(t_transform_tensor, temp_tensor);
-				for (size_t i = 0; i < dim; i++)
-					for (size_t j = 0; j < dim; j++)
-						for (size_t k = 0; k < dim; k++)
-							global_tensor(i, j) +=
-							    (*MeshElement->transform_tensor)(i, k) *
-							    temp_tensor(k, j);
-				// cout << "K:" << endl; global_tensor.Write();
-				for (size_t i = 0; i < dim; i++)
-					for (size_t j = 0; j < dim; j++)
-						k_tensor[dim * i + j] = global_tensor(i, j);
-			}
-			double const* const lambda_tensor =
-			    MediaProp->HeatDispersionTensorNew(gp, var);
-#endif
-			const double rhocp =
-			    MediaProp->HeatCapacity(Index, pcs->m_num->time_theta, this, var);
-			const double porosity =
-			    MediaProp->Porosity(Index, pcs->m_num->time_theta);
-			const double drhocp_dp =
-			    porosity * cp_w *
-			    drho_w_dp;  //  MediaProp->dHeatCapacitydP(Index,pcs->m_num->time_theta,this,var);
-			const double drhocp_dT =
-			    porosity * cp_w *
-			    drho_w_dT;  // MediaProp->dHeatCapacitydT(Index,pcs->m_num->time_theta,this,var);
-
-//---------------------------------------------------------
-//  Set velocity
-//---------------------------------------------------------
-#if 0
-		vel[0] = gp_ele->Velocity(0, gp);
-		vel[1] = gp_ele->Velocity(1, gp);
-		vel[2] = gp_ele->Velocity(2, gp);
-		if (myrank==0 && gp==0 && updateRHS)
-		ScreenMessage2("%d: vx=%.3e, vy=%.3e, vz=%.3e\n", Index, vel[0], vel[1], vel[2]);
-#else
-
-#if 1
-		for (unsigned i = 0; i < c_dim; i++)
-		{
-			vel[i] = .0;
-			for (unsigned j = 0; j < c_dim; j++)
-				vel[i] -= k_tensor[c_dim * i + j] / vis * grad_p1[j];
-			vel[i] -= k_tensor[c_dim * i + c_dim - 1] / vis * rho_w * g_const;
-		}
-#endif
-//		if (gp==0 && Index<2)
-//		ScreenMessage2("%d: vx=%.3e, vy=%.3e, vz=%.3e\n", Index, vel[0], vel[1],
-// vel[2]);
-//		if (gp==0 && Index<2)
-//		ScreenMessage2("%d: gx=%.3e, gy=%.3e, gz=%.3e\n", Index,
-// gp_ele->Velocity(0, gp), gp_ele->Velocity(1, gp), gp_ele->Velocity(2, gp));
-//		if (c_dim > MediaProp->geo_dimension) {
-//			if (gp==0 && Index%100==0)
-//				ScreenMessage2("%d: vx=%.3e, vy=%.3e, vz=%.3e\n", Index, vel[0],
-// vel[1], vel[2]);
-//			if (gp==0 && Index%100==0)
-//				ScreenMessage2("%d: gx=%.3e, gy=%.3e, gz=%.3e\n", Index,
-// gp_ele->Velocity(0, gp), gp_ele->Velocity(1, gp), gp_ele->Velocity(2, gp));
-//		}
-#endif
-
-//---------------------------------------------------------
-//  Assemble Liquid flow equation
-//  original: N^T*S*N*dp/dt + dN^T*k/mu*dN*p+dN^T*rho*g*z = 0
-//---------------------------------------------------------
-#define ENABLE_PT
-#define ENABLE_TP
-#define ENABLE_TP_MASS
-
-#if 1
-			if (updateA)
-			{
-				const double Mpp_coeff =
-				    isTransient ? dt_inverse * fkt * Ss : 0;
-				const double Lpp_coeff1 = theta * fkt / vis;
-				// const double Lpp_coeff2 = theta * fkt/(vis*vis);
-				const double Lpp_coeff3 = theta * fkt / vis * dvis_dp;
-				const double LpT_coeff3 = theta * fkt / vis * dvis_dT;
-				// const double Gpp_coeff1 = - theta * fkt / (vis*vis) * dvis_dp
-				// * rho_w * g_const;
-				const double Gpp_coeff2 =
-				    theta * fkt / vis * drho_w_dp * g_const;
-				// const double GpT_coeff3 = - theta * fkt / (vis*vis) * dvis_dT
-				// * rho_w * g_const;
-				const double GpT_coeff4 =
-				    theta * fkt / vis * drho_w_dT * g_const;
-
-				for (int i = 0; i < c_nnodes; i++)
-				{
-					// symmetric
-					for (int j = i; j < c_nnodes; j++)
-					{
-						double v_pp = .0;
-						// Jpp = 1/dt*N^T*Ss*N
-						v_pp = Mpp_coeff * c_shapefct[i] * c_shapefct[j];
-						for (unsigned k = 0; k < c_dim; k++)
-						{
-							for (unsigned l = 0; l < c_dim; l++)
-							{
-								// Jpp = theta*dN^T*k/mu*dN
-								v_pp += Lpp_coeff1 *
-								        c_dshapefct[k * c_nnodes + i] *
-								        k_tensor[c_dim * k + l] *
-								        c_dshapefct[l * c_nnodes + j];
-							}
-						}
-
-						(*StiffMatrix)(offset_p + i, offset_p + j) += v_pp;
-						if (i != j)
-						{  // make it symmetric
-							(*StiffMatrix)(offset_p + j, offset_p + i) += v_pp;
-						}
-					}
-					// unsymmetric
-					for (int j = 0; j < c_nnodes; j++)
-					{
-						double v_pp = .0;
-						double v_pT = .0;
-						for (unsigned k = 0; k < c_dim; k++)
-						{
-							const double dN_N =
-							    c_dshapefct[k * c_nnodes + i] * c_shapefct[j];
-							const double dN_kz_N =
-							    dN_N * k_tensor[c_dim * k + c_dim - 1];
-							const double dN_v_N =
-							    c_dshapefct[k * c_nnodes + i] * vel[k] *
-							    c_shapefct[j];
-							// Jpp += theta/mu*dmu_dp*dN^T*q*N
-							v_pp += Lpp_coeff3 * dN_v_N;
-							// JpT += theta/mu*dmu_dT*dN^T*q*N
-							v_pT += LpT_coeff3 * dN_v_N;
-							// Jpp += theta*dN^T*k/mu*drho_w_dp*g*N
-							v_pp += Gpp_coeff2 * dN_kz_N;
-							// Jpt += theta*dN^T*k/mu*drho_w_dT*g*N
-							v_pT += GpT_coeff4 * dN_kz_N;
-						}
-						(*StiffMatrix)(offset_p + i, offset_p + j) += v_pp;
-#ifdef ENABLE_PT
-						(*StiffMatrix)(offset_p + i, offset_T + j) += v_pT;
-#endif
-					}
-				}
-			}  // update A
-
-			if (updateRHS)
-			{
-				// Rp += [1/dt*N^T*Ss*N + theta*dN^T*k/mu*dN]*p1 -
-				// [1/dt*N^T*Ss*N - (1-theta)*dN^T*k/mu*dN]*p0 +
-				// dN^T*k/mu*rho*g*z
-				const double Rp_coeff1 =
-				    isTransient ? dt_inverse * fkt * Ss * (gp_p1 - gp_p0) : 0;
-				for (int i = 0; i < c_nnodes; i++)
-				{
-					double v_p = 0;
-					// 1/dt*N^T*Ss*(p1-p0)
-					v_p += Rp_coeff1 * c_shapefct[i];
-					for (unsigned k = 0; k < c_dim; k++)
-					{
-						// - dN^T*vel
-						v_p += -fkt * c_dshapefct[k * c_nnodes + i] * vel[k];
-					}
-					(*RHS)(offset_p + i) += v_p;
-				}
-			}  // update RHS
-               //---------------------------------------------------------
-               //  Assemble Heat transport equation
-//---------------------------------------------------------
-#ifdef CONST_HEAT
-		}
-		double varT[3] = {1750000., 333.15, .0};
-		const double rho_w = FluidProp->Density(varT);
-		const double drho_w_dp = 0;  // FluidProp->drhodP(varT);
-		const double drho_w_dT = 0;  // FluidProp->drhodT(varT);
-		const double vis = FluidProp->Viscosity(varT);
-		const double dvis_dp = 0;  // FluidProp->dViscositydP(varT);
-		const double dvis_dT = 0;  // FluidProp->dViscositydT(varT);
-		const double cp_w = FluidProp->SpecificHeatCapacity(varT);
-		const double rhocp =
-		    MediaProp->HeatCapacity(Index, pcs->m_num->time_theta, this, varT);
-		const double drhocp_dp =
-		    0;  // MediaProp->dHeatCapacitydP(Index,pcs->m_num->time_theta,this,varT);
-		const double drhocp_dT =
-		    0;  // MediaProp->dHeatCapacitydT(Index,pcs->m_num->time_theta,this,varT);
-		const int offset_p = 0;
-		const int offset_T = c_nnodes;
-#endif
-
-		double supg_tau = 0;
-		if (useSUPG) CalcSUPGWeightingFunction(vel, gp, supg_tau, weight_func);
-		double const* const c_weight_func = weight_func;
-
-		if (updateA)
-		{
-			// Jtt = 1/dt*N^T*rho*cp*N
-			// Jtt += 1/dt*N^T*drhocp_dT*T1*N
-			// Jtp += 1/dt*N^T*drhocp_dp*T1*N
-			//		const double M_tt_coeff1 = isTransient ? dt_inverse * fkt *
-			// rhocp : 0;
-			const double M_tt_coeff1 =
-			    isTransient
-			        ? dt_inverse * fkt * (rhocp + drhocp_dT * (gp_T1 - gp_T0))
-			        : 0;
-			const double M_tp_coeff1 =
-			    isTransient ? dt_inverse * fkt * drhocp_dp * (gp_T1 - gp_T0)
-			                : 0;
-			if (useLumpedMass)
-			{
-#ifdef USE_LMASS2
-				const double R_t_coeff1 =
-				    isTransient ? dt_inverse * fkt * rhocp : 0;
-				//			double vol = MeshElement->GetVolume() *
-				// MeshElement->GetFluxArea();
-				//			// Center of the reference element
-				//			SetCenterGP();
-				//			double factor = CalCoefMass();
-				//			factor *= vol / (double)nnodes;
-				//			for (int i = 0; i < nnodes; i++)
-				//				(*StiffMatrix)(offset_T+i,offset_T+i) =  factor;
-
-				for (int i = 0; i < c_nnodes; i++)
-				{
-					double NN = .0;
-					for (int j = 0; j < c_nnodes; j++)
-					{
-						NN += c_shapefct[i] * c_shapefct[j];
-					}
-					(*StiffMatrix)(offset_T + i, offset_T + i) +=
-					    M_tt_coeff1 * NN;
-#ifdef ENABLE_TP_MASS1
-					for (int k = 0; k < c_nnodes; k++)
-					{
-						//					(*StiffMatrix)(offset_T+i, offset_T+k)
-						//+=
-						// dt_inverse * fkt * drhocp_dT * NodalVal_T1[i] * NN *
-						// c_shapefct[k];
-						(*StiffMatrix)(offset_T + i, offset_p + k) +=
-						    dt_inverse * fkt * drhocp_dp * NodalVal_T1[i] * NN *
-						    c_shapefct[k];
-					}
-#endif
-					(*RHS)(offset_T + i) +=
-					    R_t_coeff1 * NN * (NodalVal_T1[i] - NodalVal_T0[i]);
-#ifdef TH_DEBUG
-					(*Mass)(i, i) += 1. / dt_inverse * R_t_coeff1 * NN;
-#endif
-				}
-#endif
-			}
-			else
-			{
-				for (int i = 0; i < c_nnodes; i++)
-					for (int j = 0; j < c_nnodes; j++)
-					{
-						const double NN = c_shapefct[i] * c_shapefct[j];
-						(*StiffMatrix)(offset_T + i, offset_T + j) +=
-						    M_tt_coeff1 * NN;
-						(*StiffMatrix)(offset_T + i, offset_p + j) +=
-						    M_tp_coeff1 * NN;
-#ifdef TH_DEBUG
-						(*Mass)(i, j) += 1. / dt_inverse * coeff_mass_T * NN;
-#endif
-					}
-				if (useSUPG)
-				{
-					// tau*({v}[dN])^T*[N]
-					for (int i = 0; i < c_nnodes; i++)
-						for (int j = 0; j < c_nnodes; j++)
-						{
-							const double NN = c_weight_func[i] * c_shapefct[j];
-							(*StiffMatrix)(offset_T + i, offset_T + j) +=
-							    supg_tau * M_tt_coeff1 * NN;
-							(*StiffMatrix)(offset_T + i, offset_p + j) +=
-							    supg_tau * M_tp_coeff1 * NN;
-#ifdef TH_DEBUG
-							(*Mass)(i, j) +=
-							    supg_tau * 1. / dt_inverse * coeff_mass_T * NN;
-#endif
-						}
-				}
-			}
-
-			// L = theta*dN^T*lambda*dN*T
-			// Jtt += theta*dN^T*lambda*dN
-			const double theta_fkt = theta * fkt;
-			const double theta_fkt_Cw = theta_fkt * rho_w * cp_w;
-			const double Att_coeff2 = theta_fkt * drho_w_dT * cp_w;
-			const double Att_coeff3 = theta_fkt * drho_w_dp * cp_w;
-			//		const double Att_coeff4 = theta_fkt_Cw / (vis*vis) *
-			// dvis_dT;
-			//		const double Att_coeff5 = theta_fkt_Cw / (vis*vis) *
-			// dvis_dp;
-			//		const double Att_coeff6 = - theta_fkt_Cw / vis;
-			//		double G_tt_coeff1 = theta_fkt_Cw / vis
-			//*(1./vis*dvis_dT*rho_w
-			//- drho_w_dT) * g_const;
-			//		double G_tp_coeff1 = theta_fkt_Cw / vis
-			//*(1./vis*dvis_dp*rho_w
-			//- drho_w_dp) * g_const;
-			for (int i = 0; i < c_nnodes; i++)
-			{
-				// symmetric
-				for (int j = i; j < c_nnodes; j++)
-				{
-					double v_TT = 0;
-					for (unsigned k = 0; k < c_dim; k++)
-					{
-						for (unsigned l = 0; l < c_dim; l++)
-						{
-							v_TT += theta_fkt * c_dshapefct[k * c_nnodes + i] *
-							        lambda_tensor[c_dim * k + l] *
-							        c_dshapefct[l * c_nnodes + j];
-#ifdef TH_DEBUG
-							(*Laplace)(i, j) += fkt *
-							                    lambda_tensor[dim * k + l] *
-							                    dshapefct[k * nnodes + i] *
-							                    dshapefct[l * nnodes + j];
-#endif
-						}
-					}
-					(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-					if (i != j)
-						(*StiffMatrix)(offset_T + j, offset_T + i) += v_TT;
-				}
-				// unsymmetric
-				for (int j = 0; j < c_nnodes; j++)
-				{
-					double v_TT = 0;
-					double v_Tp = 0;
-					const double NN = c_shapefct[i] * c_shapefct[j];
-					const double WN =
-					    useSUPG ? c_weight_func[i] * c_shapefct[j] : 0;
-
-					for (unsigned k = 0; k < c_dim; k++)
-					{
-						const double NdN =
-						    c_shapefct[i] * c_dshapefct[k * c_nnodes + j];
-						// Jtt += theta*N^T*rho_w*cp_w*q^T*dN
-						v_TT += theta_fkt_Cw * vel[k] * NdN;
-						// Jtt += theta*N^T*drho_w_dT*cp_w*q^T*grad(T)*N
-						v_TT += Att_coeff2 * vel[k] * grad_T1[k] * NN;
-						// Jtp += theta*N^T*drho_w_dp*cp_w*q^T*grad(T)*N
-						v_Tp += Att_coeff3 * vel[k] * grad_T1[k] * NN;
-#ifdef TH_DEBUG
-						(*Advection)(i, j) += fkt * shapefct[i] * rho_w * cp_w *
-						                      vel[k] *
-						                      dshapefct[k * nnodes + j];
-#endif
-						const double N_gT_kz_N =
-						    grad_T1[k] * k_tensor[c_dim * k + c_dim - 1] * NN;
-						// Jtt += -
-						// theta*N^T*rho_w*cp_w*grad(T)*k/mu*g*drho_dT*N
-						v_TT += -theta_fkt_Cw / vis * drho_w_dT * g_const *
-						        N_gT_kz_N;
-						// Jtt += - theta*N^T*rho_w*cp_w*grad(T)*1/mu*q*dmu_dT*N
-						v_TT += -theta_fkt_Cw * grad_T1[k] / vis * vel[k] *
-						        dvis_dT * NN;
-
-						// Jtp += -
-						// theta*N^T*rho_w*cp_w*grad(T)*k/mu*g*drho_dp*N
-						v_Tp += -theta_fkt_Cw / vis * drho_w_dp * g_const *
-						        N_gT_kz_N;
-						// Jtp += - theta*N^T*rho_w*cp_w*grad(T)/mu*q*dmu_dp*N
-						v_Tp += -theta_fkt_Cw * grad_T1[k] / vis * vel[k] *
-						        dvis_dp * NN;
-						for (unsigned l = 0; l < c_dim; l++)
-						{
-							// Jtp += - theta*N^T*rho_w*cp_w*grad(T)*k/mu*dN
-							v_Tp += -theta_fkt_Cw * grad_T1[k] *
-							        k_tensor[c_dim * k + l] / vis *
-							        c_shapefct[i] *
-							        c_dshapefct[l * c_nnodes + j];
-						}
-
-						if (useSUPG)
-						{
-							// Calculate mat_factor*tau*({v}[dN])^T*({v}[dN])
-							const double WdN = c_weight_func[i] *
-							                   c_dshapefct[k * c_nnodes + j];
-							v_TT += supg_tau * theta_fkt_Cw * vel[k] * WdN;
-							v_TT += supg_tau * Att_coeff2 * vel[k] *
-							        grad_T1[k] * WN;
-							v_Tp += supg_tau * Att_coeff3 * vel[k] *
-							        grad_p1[k] * WN;
-#ifdef TH_DEBUG
-							(*Advection)(i, j) +=
-							    supg_tau * fkt * rho_w * cp_w * vel[k] * WdN;
-#endif
-							const double W_gT_kz_N =
-							    grad_T1[k] * k_tensor[c_dim * k + c_dim - 1] *
-							    WN;
-							// Jtt += -
-							// theta*N^T*rho_w*cp_w*grad(T)*k/mu*g*drho_dT*N
-							v_TT += -theta_fkt_Cw / vis * drho_w_dT * g_const *
-							        W_gT_kz_N;
-							// Jtt += -
-							// theta*N^T*rho_w*cp_w*grad(T)*1/mu*q*dmu_dT*N
-							v_TT += -theta_fkt_Cw * grad_T1[k] / vis * vel[k] *
-							        dvis_dT * WN;
-
-							// Jtp += -
-							// theta*N^T*rho_w*cp_w*grad(T)*k/mu*g*drho_dp*N
-							v_Tp += -theta_fkt_Cw / vis * drho_w_dp * g_const *
-							        W_gT_kz_N;
-							// Jtp += -
-							// theta*N^T*rho_w*cp_w*grad(T)/mu*q*dmu_dp*N
-							v_Tp += -theta_fkt_Cw * grad_T1[k] / vis * vel[k] *
-							        dvis_dp * WN;
-							for (unsigned l = 0; l < c_dim; l++)
-							{
-								// Jtp += - theta*N^T*rho_w*cp_w*grad(T)*k/mu*dN
-								v_Tp += -theta_fkt_Cw * grad_T1[k] *
-								        k_tensor[c_dim * k + l] / vis *
-								        c_weight_func[i] *
-								        c_dshapefct[l * c_nnodes + j];
-							}
-						}
-					}
-#define ADV
-#ifdef ADV
-					(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-#ifdef ENABLE_TP
-					(*StiffMatrix)(offset_T + i, offset_p + j) += v_Tp;
-#endif
-#endif
-				}
-			}
-		}  // update A
-
-		if (updateRHS)
-		{
-			// Rt += [1/dt*N^T*Cp*N + theta*(dN^T*lambda*dN+N^T*Cp_w*dN)]*T1 -
-			// [1/dt*N^T*Cp*N - (1-theta)*(dN^T*lambda*dN+N^T*Cp_w*dN)]*T0
-			if (isTransient)
-			{
-				if (useLumpedMass)
-				{
-#ifdef USE_LMASS2
-					for (int i = 0; i < c_nnodes; i++)
-					{
-						double row_sum = .0;
-						for (int j = 0; j < c_nnodes; j++)
-							row_sum += dt_inverse * fkt * rhocp *
-							           c_shapefct[i] * c_shapefct[j];
-						(*RHS)(offset_T + i) +=
-						    row_sum * (NodalVal_T1[i] - NodalVal_T0[i]);
-					}
-#endif
-				}
-				else
-				{
-					const double coeff =
-					    dt_inverse * fkt * rhocp * (gp_T1 - gp_T0);
-					for (int i = 0; i < c_nnodes; i++)
-						(*RHS)(offset_T + i) += c_shapefct[i] * coeff;
-					if (useSUPG)
-					{
-						// tau*({v}[dN])^T*[N]
-						for (int i = 0; i < c_nnodes; i++)
-							(*RHS)(offset_T + i) +=
-							    supg_tau * c_weight_func[i] * coeff;
-					}
-				}
-			}
-			for (int i = 0; i < c_nnodes; i++)
-			{
-				double v_T = 0;
-				for (unsigned k = 0; k < c_dim; k++)
-				{
-					for (unsigned l = 0; l < c_dim; l++)
-					{
-						// dN^T*lambda*(theta*gradT1+(1-theta)*gradT0)
-						v_T +=
-						    fkt * c_dshapefct[k * c_nnodes + i] *
-						    lambda_tensor[c_dim * k + l] *
-						    (theta * grad_T1_r[l] + (1. - theta) * grad_T0[l]);
-					}
-#ifdef ADV
-					// N^T*Cp_w*(theta*gradT1+(1-theta)*gradT0)
-					const double cp_v_gradT =
-					    fkt * rho_w * cp_w * vel[k] *
-					    (theta * grad_T1_r[k] + (1. - theta) * grad_T0[k]);
-					v_T += c_shapefct[i] * cp_v_gradT;
-					if (useSUPG)
-					{
-						// Calculate mat_factor*tau*({v}[dN])^T*({v}[dN])
-						v_T += supg_tau * c_weight_func[i] * cp_v_gradT;
-					}
-#endif
-				}
-				(*RHS)(offset_T + i) += v_T;
-			}
-		}  // update RHS
-#else
-		const int offset_p = 0;
-		const int offset_T = c_nnodes;
-		if (this->pcs->tim_type == FiniteElement::TIM_TRANSIENT)
-		{
-			const double coeff = dt_inverse * fkt * Ss;
-			for (int i = 0; i < c_nnodes; i++)
-			{
-				(*StiffMatrix)(offset_p + i, offset_p + i) +=
-				    coeff * c_shapefct[i] * c_shapefct[i];
-				for (int j = i + 1; j < c_nnodes; j++)
-				{  // symmetric
-					// Jpp = 1/dt*N^T*Ss*N
-					const double v = coeff * c_shapefct[i] * c_shapefct[j];
-					(*StiffMatrix)(offset_p + i, offset_p + j) += v;
-					(*StiffMatrix)(offset_p + j, offset_p + i) += v;
-				}
-			}
-		}
-		const double Lpp_coeff1 = theta * fkt / vis;
-		for (int i = 0; i < c_nnodes; i++)
-		{
-			for (int j = i; j < c_nnodes; j++)
-			{  // symmetric
-				double dN_k_dN = .0;
-				for (unsigned k = 0; k < c_dim; k++)
-				{
-					for (unsigned l = 0; l < c_dim; l++)
-					{
-						// Jpp = theta*dN^T*k/mu*dN
-						dN_k_dN += c_dshapefct[k * c_nnodes + i] *
-						           k_tensor[c_dim * k + l] *
-						           c_dshapefct[l * c_nnodes + j];
-					}
-				}
-				(*StiffMatrix)(offset_p + i, offset_p + j) +=
-				    Lpp_coeff1 * dN_k_dN;
-				if (i != j)
-					(*StiffMatrix)(offset_p + j, offset_p + i) +=
-					    Lpp_coeff1 * dN_k_dN;
-			}
-		}
-		const double Lpp_coeff2 = theta * fkt / (vis * vis);
-		for (int i = 0; i < c_nnodes; i++)
-		{
-			for (int j = 0; j < c_nnodes; j++)
-			{
-				double v_pp = .0;
-				double v_pT = .0;
-				for (unsigned k = 0; k < c_dim; k++)
-				{
-					const double dN_N =
-					    c_dshapefct[k * c_nnodes + i] * c_shapefct[j];
-					double dN_k_gradP_N = .0;
-					for (unsigned l = 0; l < c_dim; l++)
-					{
-						dN_k_gradP_N +=
-						    dN_N * k_tensor[c_dim * k + l] * grad_p1[l];
-					}
-					// Jpp += -theta*dN^T*k/mu^2*dmu_dp*grad p*N
-					v_pp += -dvis_dp * Lpp_coeff2 * dN_k_gradP_N;
-					// JpT += -theta*dN^T*k/mu^2*dmu_dT*grad p*N
-					v_pT += -dvis_dT * Lpp_coeff2 * dN_k_gradP_N;
-				}
-				(*StiffMatrix)(offset_p + i, offset_p + j) += v_pp;
-				(*StiffMatrix)(offset_p + i, offset_T + j) += v_pT;
-			}
-		}
-		if (hasGravity)
-		{
-			const double coeff1 =
-			    -theta * fkt / (vis * vis) * dvis_dp * rho_w * gravity_constant;
-			const double coeff2 =
-			    theta * fkt / vis * drho_w_dp * gravity_constant;
-			const double coeff3 =
-			    -theta * fkt / (vis * vis) * dvis_dT * rho_w * gravity_constant;
-			const double coeff4 =
-			    theta * fkt / vis * drho_w_dT * gravity_constant;
-			for (int i = 0; i < c_nnodes; i++)
-				for (int j = 0; j < c_nnodes; j++)
-				{
-					double v_pp = 0;
-					double v_pT = 0;
-					for (unsigned k = 0; k < c_dim; k++)
-					{
-						const double dN_k_g_N =
-						    c_dshapefct[k * c_nnodes + i] *
-						    k_tensor[c_dim * k + c_dim - 1] * c_shapefct[j];
-						// Jpp += -theta*dN^T*k/mu^2*dmu_dp*rho_w*g*N
-						// Jpp += theta*dN^T*k/mu*drho_w_dp*g*N
-						v_pp += (coeff1 + coeff2) * dN_k_g_N;
-						// Jpt += -theta*dN^T*k/mu^2*dmu_dT*rho_w*g*N
-						// Jpt += theta*dN^T*k/mu*drho_w_dT*g*N
-						v_pT += (coeff3 + coeff4) * dN_k_g_N;
-					}
-					(*StiffMatrix)(offset_p + i, offset_p + j) += v_pp;
-					(*StiffMatrix)(offset_p + i, offset_T + j) += v_pT;
-				}
-		}
-
-		// Rp += [1/dt*N^T*Ss*N + theta*dN^T*k/mu*dN]*p1 - [1/dt*N^T*Ss*N -
-		// (1-theta)*dN^T*k/mu*dN]*p0 + dN^T*k/mu*rho*g*z
-		if (this->pcs->tim_type == FiniteElement::TIM_TRANSIENT)
-		{
-			const double coeff = dt_inverse * fkt * Ss * (gp_p1 - gp_p0);
-			for (int i = 0; i < c_nnodes; i++)
-				(*RHS)(offset_p + i) += coeff * c_shapefct[i];
-		}
-		for (int i = 0; i < c_nnodes; i++)
-		{
-			double v_p = 0;
-			for (unsigned k = 0; k < c_dim; k++)
-				for (unsigned l = 0; l < c_dim; l++)
-					v_p += fkt * c_dshapefct[k * c_nnodes + i] *
-					       k_tensor[c_dim * k + l] / vis *
-					       (theta * grad_p1[l] - (1. - theta) * grad_p0[l]);
-			(*RHS)(offset_p + i) += v_p;
-		}
-		if (hasGravity)
-		{
-			// Rp += dN^T*k/mu*rho*g
-			const double coeff = fkt / vis * rho_w * gravity_constant;
-			for (int i = 0; i < c_nnodes; i++)
-			{
-				double v_p = 0;
-				for (unsigned k = 0; k < c_dim; k++)
-					v_p += coeff * c_dshapefct[k * c_nnodes + i] *
-					       k_tensor[c_dim * k + c_dim - 1];
-				(*RHS)(offset_p + i) += v_p;
-			}
-		}
-
-		//---------------------------------------------------------
-		//  Assemble Heat transport equation
-		//---------------------------------------------------------
-		double supg_tau = 0;
-		if (pcs->m_num->ele_supg_method > 0)
-			CalcSUPGWeightingFunction(vel, gp, supg_tau, weight_func);
-
-		if (this->pcs->tim_type == FiniteElement::TIM_TRANSIENT)
-		{
-			// Jtt = 1/dt*N^T*rho*cp*N
-			// Jtt += 1/dt*N^T*drhocp_dT*T1*N
-			// Jtp += 1/dt*N^T*drhocp_dp*T1*N
-			const double coeff_mass_T = dt_inverse * fkt * rhocp;
-			const double coeff_mass_dT = dt_inverse * fkt * drhocp_dT * gp_T1;
-			const double coeff_mass_dp = dt_inverse * fkt * drhocp_dp * gp_T1;
-			if (pcs->m_num->ele_mass_lumping > 0)
-			{
-				for (int i = 0; i < c_nnodes; i++)
-				{
-					double v_TT = 0;
-					double v_Tp = 0;
-					for (int j = 0; j < c_nnodes; j++)
-					{
-						const double NN = c_shapefct[i] * c_shapefct[j];
-						v_TT += (coeff_mass_T + coeff_mass_dT) * NN;
-						v_Tp += coeff_mass_dp * NN;
-#ifdef TH_DEBUG
-						(*Mass)(i, i) += 1. / dt_inverse * coeff_mass_T * NN;
-#endif
-					}
-					(*StiffMatrix)(offset_T + i, offset_T + i) += v_TT;
-					(*StiffMatrix)(offset_T + i, offset_p + i) += v_Tp;
-				}
-			}
-			else
-			{
-				for (int i = 0; i < c_nnodes; i++)
-					for (int j = 0; j < c_nnodes; j++)
-					{
-						const double NN = c_shapefct[i] * c_shapefct[j];
-						(*StiffMatrix)(offset_T + i, offset_T + j) +=
-						    (coeff_mass_T + coeff_mass_dT) * NN;
-						(*StiffMatrix)(offset_T + i, offset_p + j) +=
-						    coeff_mass_dp * NN;
-#ifdef TH_DEBUG
-						(*Mass)(i, j) += 1. / dt_inverse * coeff_mass_T * NN;
-#endif
-					}
-				if (pcs->m_num->ele_supg_method > 0)
-				{
-					// tau*({v}[dN])^T*[N]
-					for (int i = 0; i < c_nnodes; i++)
-						for (int j = 0; j < c_nnodes; j++)
-						{
-							const double NN = weight_func[i] * c_shapefct[j];
-							(*StiffMatrix)(offset_T + i, offset_T + j) +=
-							    supg_tau * (coeff_mass_T + coeff_mass_dT) * NN;
-							(*StiffMatrix)(offset_T + i, offset_p + j) +=
-							    supg_tau * coeff_mass_dp * NN;
-#ifdef TH_DEBUG
-							(*Mass)(i, j) +=
-							    supg_tau * 1. / dt_inverse * coeff_mass_T * NN;
-#endif
-						}
-				}
-			}
-		}
-
-		// L = theta*dN^T*lambda*dN*T
-		// Jtt += theta*dN^T*lambda*dN
-		const double Ltt_coeff = theta * fkt;
-		for (int i = 0; i < c_nnodes; i++)
-		{
-			for (int j = i; j < c_nnodes; j++)
-			{  // symmetric
-				double v_TT = 0;
-				for (unsigned k = 0; k < c_dim; k++)
-				{
-					for (unsigned l = 0; l < c_dim; l++)
-					{
-						v_TT += Ltt_coeff * c_dshapefct[k * c_nnodes + i] *
-						        lambda_tensor[c_dim * k + l] *
-						        c_dshapefct[l * c_nnodes + j];
-#ifdef TH_DEBUG
-						(*Laplace)(i, j) += fkt * lambda_tensor[dim * k + l] *
-						                    dshapefct[k * nnodes + i] *
-						                    dshapefct[l * nnodes + j];
-#endif
-					}
-				}
-				(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-				if (i != j) (*StiffMatrix)(offset_T + j, offset_T + i) += v_TT;
-			}
-		}
-
-		// A = theta*N^T*rho_w*cp_w*q^T*dN*T
-		const double Att_coeff1 = theta * fkt * rho_w * cp_w;
-		const double Att_coeff2 = theta * fkt * drho_w_dT * cp_w;
-		const double Att_coeff3 = theta * fkt * drho_w_dp * cp_w;
-		const double Att_coeff4 =
-		    theta * fkt * rho_w * cp_w / (vis * vis) * dvis_dT;
-		const double Att_coeff5 =
-		    theta * fkt * rho_w * cp_w / (vis * vis) * dvis_dp;
-		const double Att_coeff6 = -theta * fkt * rho_w * cp_w / vis;
-		for (int i = 0; i < c_nnodes; i++)
-		{
-			for (int j = 0; j < c_nnodes; j++)
-			{
-				double v_TT = 0;
-				double v_Tp = 0;
-				const double NN = c_shapefct[i] * c_shapefct[j];
-				for (unsigned k = 0; k < c_dim; k++)
-				{
-					const double NdN =
-					    c_shapefct[i] * c_dshapefct[k * c_nnodes + j];
-					// Jtt += theta*N^T*rho_w*cp_w*q^T*dN
-					v_TT += Att_coeff1 * vel[k] * NdN;
-					// Jtt += theta*N^T*drho_w_dT*cp_w*q^T*grad(T)*N
-					v_TT += Att_coeff2 * vel[k] * grad_T1[k] * NN;
-					// Jtp += theta*N^T*drho_w_dp*cp_w*q^T*grad(T)*N
-					v_Tp += Att_coeff3 * vel[k] * grad_T1[k] * NN;
-#ifdef TH_DEBUG
-					(*Advection)(i, j) += fkt * shapefct[i] * rho_w * cp_w *
-					                      vel[k] * dshapefct[k * nnodes + j];
-#endif
-					for (unsigned l = 0; l < c_dim; l++)
-					{
-						const double gT_k_gp_NN = grad_T1[k] *
-						                          k_tensor[c_dim * k + l] *
-						                          grad_p1[l] * NN;
-						// Jtt +=
-						// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu^2*dmu_dT*grad(p)*N
-						v_TT += Att_coeff4 * gT_k_gp_NN;
-						// Jtp +=
-						// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu^2*dmu_dp*grad(p)*N
-						v_Tp += Att_coeff5 * gT_k_gp_NN;
-// Jtp += theta*N^T*rho_w*cp_w*grad(T)^T*(-k/mu)*dN
-#if 0
-						v_Tp += Att_coeff6 * grad_T1[k] * k_tensor[c_dim * k + l] * NdN;
-#else
-						v_Tp += Att_coeff6 * c_shapefct[i] * grad_T1[k] *
-						        k_tensor[c_dim * k + l] *
-						        c_dshapefct[l * c_nnodes + j];
-#endif
-					}
-				}
-				(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-				(*StiffMatrix)(offset_T + i, offset_p + j) += v_Tp;
-			}
-		}
-		if (hasGravity)
-		{
-			double coeff1 = theta * fkt * rho_w * cp_w / vis *
-			                (1. / vis * dvis_dT * rho_w - drho_w_dT) *
-			                gravity_constant;
-			double coeff2 = theta * fkt * rho_w * cp_w / vis *
-			                (1. / vis * dvis_dp * rho_w - drho_w_dp) *
-			                gravity_constant;
-			for (int i = 0; i < c_nnodes; i++)
-				for (int j = 0; j < c_nnodes; j++)
-				{
-					const double NN = c_shapefct[i] * c_shapefct[j];
-					double v_TT = .0;
-					double v_Tp = .0;
-					for (unsigned k = 0; k < c_dim; k++)
-					{
-						const double N_gT_k_N =
-						    grad_T1[k] * k_tensor[c_dim * k + c_dim - 1] * NN;
-						// Jtt +=
-						// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu*[-1/mu*dmu_dT*rho+drho_dT]*g*N
-						v_TT += coeff1 * N_gT_k_N;
-						// Jtp +=
-						// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu*[-1/mu*dmu_dp*rho+drho_dp]*g*N
-						v_Tp += coeff2 * N_gT_k_N;
-					}
-					(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-					(*StiffMatrix)(offset_T + i, offset_p + j) += v_Tp;
-				}
-		}
-
-		if (pcs->m_num->ele_supg_method > 0)
-		{
-			// Calculate mat_factor*tau*({v}[dN])^T*({v}[dN])
-			for (int i = 0; i < c_nnodes; i++)
-				for (int j = 0; j < c_nnodes; j++)
-				{
-					const double NN = weight_func[i] * c_shapefct[j];
-					double v_TT = 0;
-					double v_Tp = 0;
-					for (unsigned k = 0; k < c_dim; k++)
-					{
-						const double NdN =
-						    weight_func[i] * c_dshapefct[k * c_nnodes + j];
-						v_TT += supg_tau * Att_coeff1 * vel[k] * NdN;
-						v_TT +=
-						    supg_tau * Att_coeff2 * vel[k] * grad_T1[k] * NN;
-						v_Tp +=
-						    supg_tau * Att_coeff3 * vel[k] * grad_p1[k] * NN;
-#ifdef TH_DEBUG
-						(*Advection)(i, j) +=
-						    supg_tau * fkt * rho_w * cp_w * vel[k] * NdN;
-#endif
-						for (unsigned l = 0; l < c_dim; l++)
-						{
-							const double gT_k_gp_NN = grad_T1[k] *
-							                          k_tensor[c_dim * k + l] *
-							                          grad_p1[l] * NN;
-							// Jtt +=
-							// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu^2*dmu_dT*grad(p)*N
-							v_TT += supg_tau * Att_coeff4 * gT_k_gp_NN;
-							// Jtp +=
-							// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu^2*dmu_dp*grad(p)*N
-							v_Tp += supg_tau * Att_coeff5 * gT_k_gp_NN;
-// Jtp += theta*N^T*rho_w*cp_w*grad(T)^T*(-k/mu)*dN
-#if 0
-							v_Tp += supg_tau * Att_coeff6 * grad_T1[k] * k_tensor[c_dim * k + l] * NdN;
-#else
-							v_Tp += supg_tau * Att_coeff6 * weight_func[i] *
-							        grad_T1[k] * k_tensor[c_dim * k + l] *
-							        c_dshapefct[l * c_nnodes + j];
-#endif
-						}
-					}
-					(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-					(*StiffMatrix)(offset_T + i, offset_p + j) += v_Tp;
-				}
-
-			if (hasGravity)
-			{
-				double coeff1 = theta * fkt * rho_w * cp_w / vis *
-				                (1. / vis * dvis_dT * rho_w - drho_w_dT) *
-				                gravity_constant;
-				double coeff2 = theta * fkt * rho_w * cp_w / vis *
-				                (1. / vis * dvis_dp * rho_w - drho_w_dp) *
-				                gravity_constant;
-				for (int i = 0; i < c_nnodes; i++)
-					for (int j = 0; j < c_nnodes; j++)
-					{
-						const double NN = weight_func[i] * c_shapefct[j];
-						double v_TT = 0;
-						double v_Tp = 0;
-						for (unsigned k = 0; k < c_dim; k++)
-						{
-							const double N_gT_k_N =
-							    grad_T1[k] * k_tensor[c_dim * k + c_dim - 1] *
-							    NN;
-							// Jtt +=
-							// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu*[-1/mu*dmu_dT*rho+drho_dT]*g*N
-							v_TT += supg_tau * coeff1 * N_gT_k_N;
-							// Jtp +=
-							// theta*N^T*rho_w*cp_w*grad(T)^T*k/mu*[-1/mu*dmu_dp*rho+drho_dp]*g*N
-							v_Tp += supg_tau * coeff2 * N_gT_k_N;
-						}
-						(*StiffMatrix)(offset_T + i, offset_T + j) += v_TT;
-						(*StiffMatrix)(offset_T + i, offset_p + j) += v_Tp;
-					}
-			}
-		}
-
-		// Rt += [1/dt*N^T*Cp*N + theta*(dN^T*lambda*dN+N^T*Cp_w*dN)]*T1 -
-		// [1/dt*N^T*Cp*N - (1-theta)*(dN^T*lambda*dN+N^T*Cp_w*dN)]*T0
-		if (this->pcs->tim_type == FiniteElement::TIM_TRANSIENT)
-		{
-			if (pcs->m_num->ele_mass_lumping > 0)
-			{
-				for (int i = 0; i < c_nnodes; i++)
-				{
-					double row_sum = .0;
-					for (int j = 0; j < c_nnodes; j++)
-						row_sum += dt_inverse * fkt * rhocp * c_shapefct[i] *
-						           c_shapefct[j];
-					(*RHS)(offset_T + i) +=
-					    row_sum * (NodalVal_T1[i] - NodalVal_T0[i]);
-				}
-			}
-			else
-			{
-				const double coeff = dt_inverse * fkt * rhocp * (gp_T1 - gp_T0);
-				for (int i = 0; i < c_nnodes; i++)
-					(*RHS)(offset_T + i) += c_shapefct[i] * coeff;
-				if (pcs->m_num->ele_supg_method > 0)
-				{
-					// tau*({v}[dN])^T*[N]
-					for (int i = 0; i < c_nnodes; i++)
-						(*RHS)(offset_T + i) +=
-						    supg_tau * weight_func[i] * coeff;
-				}
-			}
-		}
-		for (int i = 0; i < c_nnodes; i++)
-		{
-			double v_T = 0;
-			for (unsigned k = 0; k < c_dim; k++)
-				for (unsigned l = 0; l < c_dim; l++)
-					v_T += fkt * lambda_tensor[c_dim * k + l] *
-					       c_dshapefct[k * c_nnodes + i] *
-					       (theta * grad_T1[l] - (1. - theta) * grad_T0[l]);
-			(*RHS)(offset_T + i) += v_T;
-		}
-		for (int i = 0; i < c_nnodes; i++)
-			for (size_t k = 0; k < c_dim; k++)
-				(*RHS)(offset_T + i) +=
-				    fkt * c_shapefct[i] * rho_w * cp_w * vel[k] *
-				    (theta * grad_T1[k] - (1. - theta) * grad_T0[k]);
-		if (pcs->m_num->ele_supg_method > 0)
-		{
-			// Calculate mat_factor*tau*({v}[dN])^T*({v}[dN])
-			for (int i = 0; i < c_nnodes; i++)
-				for (unsigned k = 0; k < c_dim; k++)
-					(*RHS)(offset_T + i) +=
-					    supg_tau * fkt * weight_func[i] * rho_w * cp_w *
-					    vel[k] *
-					    (theta * grad_T1[k] - (1. - theta) * grad_T0[k]);
-		}
-#endif
-	}
-
-	// RHS should be - residual
-	if (updateRHS) (*RHS) *= -1.;
-
-#if !defined(USE_PETSC)
-	for (size_t ii = 0; ii < pcs->GetPrimaryVNumber(); ii++)
-		for (long i = 0; i < c_nnodes; i++)
-			eqs_rhs[NodeShift[ii] + eqs_number[i]] += (*RHS)(i + ii * c_nnodes);
-#endif
-
-	if (updateA && (pcs->scaleUnknowns || pcs->scaleEQS))
-	{
-		// scale p by 1e6
-		const double scale_p =
-		    1. / pcs->vec_scale_dofs[0] * pcs->vec_scale_eqs[0];
-		for (long i = 0; i < c_nnodes; i++)
-			for (long j = 0; j < c_nnodes; j++)
-			{
-				(*StiffMatrix)(i, j) *= scale_p;
-				(*StiffMatrix)(i, j + c_nnodes) *= scale_p;
-				//				(*StiffMatrix)(i+c_nnodes,j) *= scale_p;
-			}
-		// scale T
-		const double scale_T =
-		    1. / pcs->vec_scale_dofs[1] * pcs->vec_scale_eqs[1];
-		for (long i = 0; i < c_nnodes; i++)
-			for (long j = 0; j < c_nnodes; j++)
-			{
-				(*StiffMatrix)(i + c_nnodes, j) *= scale_T;
-				(*StiffMatrix)(i + c_nnodes, j + c_nnodes) *= scale_T;
-				//				(*StiffMatrix)(i,j+c_nnodes) *= scale_T;
-				//				(*StiffMatrix)(i+c_nnodes,j+c_nnodes) *=
-				// scale_T;
-			}
-	}
-	if (updateRHS && pcs->scaleEQS)
-	{
-		for (int ii = 0; ii < 2; ii++)
-		{
-			const double scale_eqs = pcs->vec_scale_eqs[ii];
-			for (int i = 0; i < c_nnodes; i++)
-				(*RHS)(ii* c_nnodes + i) *= scale_eqs;
-		}
-	}
-}
-
 }  // end namespace
 
-//////////////////////////////////////////////////////////////////////////
-
-using FiniteElement::ElementValue;
-vector<ElementValue*> ele_gp_value;
